% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{add_idvars}
\alias{add_idvars}
\title{Add future, historical identifiers to scens object}
\usage{
add_idvars(
  scens,
  full,
  confirm = TRUE,
  overwrite = TRUE,
  save = TRUE,
  ignores = c("futCMcov", "futITNcov2022", "futITNcov2023")
)
}
\arguments{
\item{scens}{Scens object}

\item{full}{List of experiment variables}

\item{confirm}{Prints message if TRUE}

\item{overwrite}{Overwrites existing fut, HistScen_nr}

\item{save}{Saves the "param_names.RDS" file}

\item{ignores}{Variables to ignore}
}
\description{
Add future, historical identifiers to scens object
}
