% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_demography.R
\name{ageGroupsGen}
\alias{ageGroupsGen}
\title{Generate list for 'demography/ageGroup'}
\usage{
ageGroupsGen(lowerbound, ageGroups)
}
\arguments{
\item{lowerbound}{Double, lower bound of age group}

\item{ageGroups}{Data frame containing demography}
}
\value{
List for xml contruction
}
\description{
Generate list for 'demography/ageGroup'
}
