% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{assign_value}
\alias{assign_value}
\title{Replaces text for a 'level' with a numeric 'value'}
\usage{
assign_value(variable = "futIRScov", levels = "none", values = 0, scens)
}
\arguments{
\item{variable}{futIRScov}

\item{levels}{c('none','high')}

\item{values}{c(0, .85 )}

\item{scens}{scens object}
}
\description{
Replaces text for a 'level' with a numeric 'value'
}
\examples{
scens <- data.frame(
  setting = "alpha", futIRScov = c("none", "curr"),
  futITNcov = c("none", "high")
)
scens <- assign_value(
  variable = "futIRScov", levels = c("none", "curr"),
  values = c(0, .5), scens = scens
)
scens <- assign_value(
  variable = "futITNcov", levels = c("none", "high"),
  values = c(0, .8), scens = scens
)
}
