% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BEN}
\alias{BEN}
\title{Demographic data from Benin}
\format{
An object of class \code{data.frame} with 18 rows and 2 columns.
}
\usage{
BEN
}
\description{
A dataset with demographic data from Benin. The variables are as follows:
}
\keyword{datasets}
