% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess-new.R
\name{calculateEpidemiologicalIndicators}
\alias{calculateEpidemiologicalIndicators}
\title{Calculate epidemiological indicators from stacked raw data}
\usage{
calculateEpidemiologicalIndicators(
  rawdata = NULL,
  metadata = NULL,
  metadataFeatures = NULL,
  indicators = c("incidence", "prevalence"),
  keepMeasuresForIndicators = NULL
)
}
\arguments{
\item{rawdata}{Stacked survey output data file location, by default stackedSurveyOutput.dat file}

\item{metadata}{Experiment metadata file location, by default scens object from cache}

\item{metadataFeatures}{Character vector of metadata features to keep, if null keep all}

\item{indicators}{Character vector of epidemiological indicators to calculate}

\item{keepMeasures}{Character vector of indicators whose measures you want to keep in the function output, if null keep none}
}
\description{
Calculate epidemiological indicators from stacked raw data
}
