% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_interventions_changeHS.R
\name{changeHSCFRGen}
\alias{changeHSCFRGen}
\title{Generate list for 'changeHS/CFR'}
\usage{
changeHSCFRGen(interpolation = NULL, ageGroups)
}
\arguments{
\item{interpolation}{Value for interpolation}

\item{ageGroups}{Data frame}
}
\description{
Generate list for 'changeHS/CFR'
}
