% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_interventions_changeHS.R
\name{changeHSpSeqInGen}
\alias{changeHSpSeqInGen}
\title{TODO}
\usage{
changeHSpSeqInGen(interpolation = NULL, ageGroups)
}
\arguments{
\item{interpolation}{TODO}

\item{ageGroups}{TODO}
}
\value{
TODO
}
\description{
TODO
}
