% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logs.R
\name{cleanLogs}
\alias{cleanLogs}
\alias{clean_logs}
\title{Clean log directories}
\usage{
cleanLogs(aggregate = TRUE, compress = TRUE)

clean_logs(aggregate = TRUE, compress = TRUE)
}
\arguments{
\item{aggregate}{If TRUE, aggregate all files into a single one.}

\item{compress}{If TRUE, use create a zip archive per directory.}
}
\description{
Clean log directories
}
