% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cleanupExperiment}
\alias{cleanupExperiment}
\title{Cleanup an experiment}
\usage{
cleanupExperiment()
}
\description{
Cleaning up includes the aggregation and compression of the log
files, as well as the XML files in the scenarios folder and the output
files of Open Malaria.
}
