% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CMR}
\alias{CMR}
\title{Demographic data from Cameroon}
\format{
An object of class \code{data.frame} with 20 rows and 2 columns.
}
\usage{
CMR
}
\description{
A dataset with demographic data from Cameroon. The variables are as follows:
}
\keyword{datasets}
