% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{convert_cm}
\alias{convert_cm}
\title{Convert access to effective treatment coverage}
\usage{
convert_cm(orig, katya = FALSE, country = NULL, scale = NULL, reverse = F)
}
\arguments{
\item{orig}{Original values of access to care}

\item{katya}{Convert from access to effective treatment coverage (boolean)}

\item{country}{Abbreviation of the country, if NULL default will be used}

\item{scale}{Scaling factor (integer), if NULL, a default country specific
scaling factor will be used}

\item{reverse}{Reverse the conversion (default from access to effective
treatment coverage), (boolean)}
}
\description{
Convert access to effective treatment coverage
}
\examples{
# Converting 80\% access to care to a 5-day OpenMalaria timestep
convert_cm(orig = .80, scale = 1)

# Converting 80\% access to care to a 5-day OpenMalaria timestep
# in Ghana, where it is assumed that 63.7\% are effectively treated
convert_cm(orig = .80, country = "GHA")

# This is equivalent to the following
convert_cm(orig = .80 * .637, scale = 1)

# Going from OpenMalaria timesteps to access to care is also possible
convert_cm(orig = .249, reverse = TRUE, country = "GHA")
}
