% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_interventions_changeHS.R
\name{define_changeHS_compat}
\alias{define_changeHS_compat}
\title{Function to write the case management deployments}
\usage{
define_changeHS_compat(
  baseList,
  access = "Access",
  coverage = NULL,
  y1 = 2000,
  y2 = 2015,
  use_at_symbol = TRUE,
  pSelfTreatUncomplicated = 0.01821375,
  futSevere = NULL,
  pSeekOfficialCareSevere = 0.48,
  SIMSTART = "1918-01-01",
  every = 1,
  interval = "year",
  m1 = 1,
  m2 = 1,
  init.act = 1,
  init.qn = 1,
  init.self = 1,
  comp.act = 1,
  comp.qn = 1,
  comp.self = 1
)
}
\arguments{
\item{baseList}{List with experiment data.}

\item{access}{Name of coverage value (i.e. "Access")}

\item{coverage}{Name of future intervention for change in HS (i.e.
"@futCM@")}

\item{y1}{Year of the first date (surveys starting from year y1)}

\item{y2}{Year of the end date (surveys continuing until year y2)}

\item{use_at_symbol}{If TRUE, variable will be replaced with setting
specific values}

\item{pSelfTreatUncomplicated}{Proportion of uncomplicated episodes treated
at home}

\item{futSevere}{Variable for different values of 'pseekOfficialCareSevere'}

\item{pSeekOfficialCareSevere}{Proportion of severe episodes that seek care
at the formal sector}

\item{SIMSTART}{Start of the simulations (equal to ORIGIN!)}

\item{every}{Interval size}

\item{interval}{Interval size (days, weeks, )}

\item{m1}{Month of the first date}

\item{m2}{Month of the end date}

\item{init.act}{Initial ACT}

\item{init.qn}{Initial Quinine}

\item{init.self}{Initial self treatment}

\item{comp.act}{See OpenMalaria}

\item{comp.qn}{See OpenMalaria}

\item{comp.self}{See OpenMalaria}
}
\description{
Function to write the case management deployments
}
