% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_entomology_define.R
\name{define_entomology}
\alias{define_entomology}
\title{Defines entomology setting in baseXMLfile}
\usage{
define_entomology(baseXMLfile, mosquito_parameters, append = T)
}
\arguments{
\item{baseXMLfile}{xml file to which this functions appends to}

\item{mosquito_parameters}{mosquito bionomics parameterization list of different mosquito species as obtained from AnophelesModel package function get_OM_ento_snippet}

\item{append}{if T, then append to existing baseXMLfile, otherwise overwrite}
}
\description{
Defines entomology setting in baseXMLfile
}
