% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_interventions_define.R
\name{define_importedInfections_compat}
\alias{define_importedInfections_compat}
\title{Writes importation intervention}
\usage{
define_importedInfections_compat(baseList, val = 10, time = 0)
}
\arguments{
\item{baseList}{List with experiment data.}

\item{val}{Number of imported infections per 1000.}

\item{time}{Rate of importation, if 0, constant importation rate since the
beginning.}
}
\description{
Writes importation intervention
}
