% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_interventions_define.R
\name{define_IRS_compat}
\alias{define_IRS_compat}
\title{Adds the IRS intervention parameterisation. Compatibility version.}
\usage{
define_IRS_compat(
  baseList,
  mosqs,
  component = "Actellic50EC",
  noeffect = "outdoor",
  steplife = NULL
)
}
\arguments{
\item{baseList}{List with experiment data.}

\item{mosqs}{Mosquito species affected by the intervention.}

\item{component}{Insecticde to use. This will automatically specify efficacy
and effect duration. Needs to from predefined choices.}

\item{noeffect}{Pattern which defines which mosquitoes are unaffected by
intervention. Can be "outdoor", "indoor" or NULL.}

\item{steplife}{Step function decay, the number of months it's effective.}
}
\description{
Adds the IRS intervention parameterisation. Compatibility version.
}
