% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_interventions_define.R
\name{define_ITN_compat}
\alias{define_ITN_compat}
\title{Writes the ITN intervention parameterisation xml chunk. Compatibility
version.}
\usage{
define_ITN_compat(
  baseList,
  component = "histITN",
  noeffect = "outdoor",
  mosqs,
  halflife = 2,
  resist = TRUE,
  hist = FALSE,
  strong = FALSE,
  versionnum = 38
)
}
\arguments{
\item{baseList}{List with experiment data.}

\item{component}{Name of the intervention, can be any name but needs to be
the same as defined in deployment.}

\item{noeffect}{Which mosquitoes unaffected by intervention?}

\item{mosqs}{Name of mosquito species affected by the intervention}

\item{halflife}{Attrition of nets in years}

\item{resist}{If TRUE, a pyrethroid resistance will be assumed (default
percentage..?)}

\item{hist}{Used for historical intervention coverage?}

\item{strong}{If !strong and !resist, then "Pitoa" parameter for LLIN}

\item{versionnum}{OpenMalaria version (e.g. 38, 43)}
}
\description{
Writes the ITN intervention parameterisation xml chunk. Compatibility
version.
}
