% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_interventions_define.R
\name{define_larv_compat}
\alias{define_larv_compat}
\title{Function that writes the larviciding component xml chunk}
\usage{
define_larv_compat(
  baseList,
  mosqs,
  component = "LSM",
  coverage = "@futLSMcov@",
  decayVals = list(L = 0.25, k = NULL, funct = "step"),
  y1 = 2018,
  y2 = 2030,
  m1 = 4,
  m2 = 6,
  every = 2,
  interval = "week",
  SIMSTART = "1918-01-01"
)
}
\arguments{
\item{baseList}{List with experiment data.}

\item{mosqs}{Names of mosquito species to simulate}

\item{component}{Name of vector population intervention (i.e. LSM)}

\item{coverage}{A variable or a numeric value}

\item{decayVals}{Decay (if not specified, constant effectiveness, step
function as default)}

\item{y1}{Year of the first date (surveys starting from year y1)}

\item{y2}{Year of the end date (surveys continuing until year y2)}

\item{m1}{Month of the first date}

\item{m2}{Month of the end date}

\item{every}{Integer value}

\item{interval}{Month, year, week, quarter}

\item{SIMSTART}{Starting date of the simulations in the format "yyyy-mm-dd",
default e.g. "1918-01-01"}
}
\description{
Function that writes the larviciding component xml chunk
}
