% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_interventions_define.R
\name{define_nothing_compat}
\alias{define_nothing_compat}
\title{Writes an intervention parameterisation xml chunk that does nothing.
Compatibility version.}
\usage{
define_nothing_compat(baseList, mosqs, component = "nothing")
}
\arguments{
\item{baseList}{List with experiment data.}

\item{mosqs}{Name of mosquito species affected by the intervention.}

\item{component}{Name of the intervention, can be any name but needs to be
the same as defined in deployment}
}
\description{
This is useful if something needs to be deployed, as a
placeholder.
}
