% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_interventions_define.R
\name{define_vaccine}
\alias{define_vaccine}
\title{Adds vaccine intervention parameterisation to baseXMLfile}
\usage{
define_vaccine(
  baseXMLfile,
  vaccine_parameterization,
  append = T,
  name = NULL,
  hist = F
)
}
\arguments{
\item{baseXMLfile}{xml file to which this functions appends to}

\item{append}{if T, then append to existing baseXMLfile, otherwise overwrites}

\item{name}{name tag list}

\item{hist}{if T, then decay is assumed to be step function set to 1 for a year and then to zero for the remainder TODO}

\item{vaccine}{parameterization list (see example below)}
}
\description{
Adds vaccine intervention parameterisation to baseXMLfile
}
