% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_interventions_define.R
\name{define_vector_control}
\alias{define_vector_control}
\title{Adds vector control intervention parameterisation to baseXMLfile}
\usage{
define_vector_control(
  baseXMLfile,
  intervention_parameterization,
  append = T,
  name = NULL,
  hist = F,
  resistance = 0.1
)
}
\arguments{
\item{baseXMLfile}{xml file to which this functions appends to}

\item{intervention_parameterization}{vector control intervention parameterization list depending on three parameters (deterrency, preprandrial, postprandial) and decay functions:}

\item{append}{if T, then append to existing baseXMLfile, otherwise overwrites}

\item{name}{name tag list}

\item{hist}{if T, then decay is assumed to be step function set to 1 for a year and then to zero for the remainder}

\item{resistance}{scaling function of insecticide resistance ##TODO}
}
\description{
Adds vector control intervention parameterisation to baseXMLfile
}
