% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_interventions_changeHS.R
\name{defineChangeHS}
\alias{defineChangeHS}
\alias{define_changeHS}
\title{Write the case management deployments}
\usage{
defineChangeHS(
  baseList,
  name = "Change in case management",
  startDate = NULL,
  endDate = NULL,
  interval = NULL,
  dates = NULL,
  initACT = 1,
  initQN = 1,
  initSelf = 1,
  compACT = 1,
  compQN = 1,
  compSelf = 1,
  pSeekOfficialCareUncomplicated1,
  pSelfTreatUncomplicated = 0.01821375,
  pSeekOfficialCareUncomplicated2,
  pSeekOfficialCareSevere = 0.48
)

define_changeHS(
  baseList,
  name = "Change in case management",
  startDate = NULL,
  endDate = NULL,
  interval = NULL,
  dates = NULL,
  initACT = 1,
  initQN = 1,
  initSelf = 1,
  compACT = 1,
  compQN = 1,
  compSelf = 1,
  pSeekOfficialCareUncomplicated1,
  pSelfTreatUncomplicated = 0.01821375,
  pSeekOfficialCareUncomplicated2,
  pSeekOfficialCareSevere = 0.48
)
}
\arguments{
\item{baseList}{List with experiment data.}

\item{name}{Name of the intervention}

\item{startDate}{Date in YYYY-MM-DD format.}

\item{endDate}{Date in YYYY-MM-DD format.}

\item{interval}{A string like '1 weeks'. Same as in \code{\link[=seq.Date]{seq.Date()}}. Or a list
composed of the entries 'days' (optional), 'months' (optional) and
'years'. If a list is used, startDate and endDate are not used and can be
NULL.}

\item{dates}{If NULL, startDate, endDate and interval are used, else a
vector of dates in YYYY-MM-DD format. Can be a placeholder.}

\item{initACT}{Initial artemisinine combination therapy. Can be a placeholder.}

\item{initQN}{Initial quinine. Can be a placeholder.}

\item{initSelf}{Initial probability of self-treatment. Can be a placeholder.}

\item{compACT}{Compliance artemisinine combination therapy. Can be a placeholder.}

\item{compQN}{Compliance quinine. Can be a placeholder.}

\item{compSelf}{Compliance to self-treatment. Can be a placeholder.}

\item{pSeekOfficialCareUncomplicated1}{Probability that a patient with newly
incident uncomplicated disease seeks official care. Can be a placeholder.}

\item{pSelfTreatUncomplicated}{Probability that a patient with uncomplicated
disease without recent history of disease (i.e. first line) will
self-treat. Can be a placeholder.}

\item{pSeekOfficialCareUncomplicated2}{Probability that a patient with
recurrence of uncomplicated disease seeks official care. Can be a placeholder.}

\item{pSeekOfficialCareSevere}{Probability that a patient with severe
disease obtains appropriate care. Can be a placeholder.}
}
\description{
Write the case management deployments
}
