% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_interventions_define.R
\name{defineCohort}
\alias{defineCohort}
\alias{define_cohort}
\title{Writes cohort recruitment-only component XML chunks}
\usage{
defineCohort(baseList, ids)

define_cohort(baseList, ids)
}
\arguments{
\item{baseList}{List with experiment data.}

\item{ids}{Character vector of cohort component identifiers.}
}
\description{
Writes cohort recruitment-only component XML chunks
}
