% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_demography.R
\name{defineDemography}
\alias{defineDemography}
\alias{define_demography}
\title{Define and write demography input to baseList}
\usage{
defineDemography(
  baseList,
  name,
  popSize = 3000,
  maximumAgeYrs,
  growthRate = NULL,
  lowerbound,
  poppercent,
  upperbound
)

define_demography(
  baseList,
  name,
  popSize = 3000,
  maximumAgeYrs,
  growthRate = NULL,
  lowerbound,
  poppercent,
  upperbound
)
}
\arguments{
\item{baseList}{List with experiment data.}

\item{name}{Name of demography data.}

\item{popSize}{Population size.}

\item{maximumAgeYrs}{Maximum age of simulated humans in years.}

\item{growthRate}{Growth rate of human population.}

\item{lowerbound}{Lower bound of age group.}

\item{poppercent}{Percentage of human population in age group. Must be a
numerical vector.}

\item{upperbound}{Upper bound of age group. Must be a numerical vector.}
}
\description{
Define and write demography input to baseList
}
