% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_healthSystem.R
\name{defineHealthSystem}
\alias{defineHealthSystem}
\alias{define_health_system}
\title{Writes the health system part}
\usage{
defineHealthSystem(
  baseList,
  name = "Tanzania ACT",
  initACT = 1,
  initQN = 1,
  initSelf = 1,
  compACT = 1,
  compQN = 1,
  compSelf = 1,
  pSeekOfficialCareUncomplicated1,
  pSelfTreatUncomplicated = 0.01821375,
  pSeekOfficialCareUncomplicated2,
  pSeekOfficialCareSevere = 0.48
)

define_health_system(
  baseList,
  name = "Tanzania ACT",
  initACT = 1,
  initQN = 1,
  initSelf = 1,
  compACT = 1,
  compQN = 1,
  compSelf = 1,
  pSeekOfficialCareUncomplicated1,
  pSelfTreatUncomplicated = 0.01821375,
  pSeekOfficialCareUncomplicated2,
  pSeekOfficialCareSevere = 0.48
)
}
\arguments{
\item{baseList}{List with experiment data.}

\item{name}{Name of health system}

\item{initACT}{Initial artemisinine combination therapy.}

\item{initQN}{Initial quinine.}

\item{initSelf}{Initial probability of self-treatment.}

\item{compACT}{Compliance artemisinine combination therapy.}

\item{compQN}{Compliance quinine.}

\item{compSelf}{Compliance to self-treatment.}

\item{pSeekOfficialCareUncomplicated1}{Probability that a patient with newly
incident uncomplicated disease seeks official care.}

\item{pSelfTreatUncomplicated}{Probability that a patient with uncomplicated
disease without recent history of disease (i.e. first line) will
self-treat.}

\item{pSeekOfficialCareUncomplicated2}{Probability that a patient with
recurrence of uncomplicated disease seeks official care.}

\item{pSeekOfficialCareSevere}{Probability that a patient with severe
disease obtains appropriate care.}
}
\description{
Writes the health system part
}
