% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_interventions_define.R
\name{defineImportedInfections}
\alias{defineImportedInfections}
\alias{define_importedInfections}
\title{Writes importation intervention}
\usage{
defineImportedInfections(
  baseList,
  name = "importedInfections",
  value = 10,
  time = 0
)

define_importedInfections(
  baseList,
  name = "importedInfections",
  value = 10,
  time = 0
)
}
\arguments{
\item{baseList}{List with experiment data.}

\item{name}{Name of the intervention.}

\item{value}{Number of imported infections per 1000.}

\item{time}{Rate of importation, if 0, constant importation rate since the
beginning.}
}
\description{
Models importation of P. falciparum infections directly into
humans from an external source. This is infections, not inoculations or
EIR being imported.
}
