% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_interventions_define.R
\name{defineLarv}
\alias{defineLarv}
\alias{define_larv}
\title{Function that writes the larviciding component}
\usage{
defineLarv(
  baseList,
  mosquitos,
  component = "LSM",
  coverage = "@futLSMcov@",
  decayVals = list(L = 0.25, k = NULL, funct = "step"),
  startDate = NULL,
  endDate = NULL,
  interval,
  dates = NULL
)

define_larv(
  baseList,
  mosquitos,
  component = "LSM",
  coverage = "@futLSMcov@",
  decayVals = list(L = 0.25, k = NULL, funct = "step"),
  startDate = NULL,
  endDate = NULL,
  interval,
  dates = NULL
)
}
\arguments{
\item{baseList}{List with experiment data.}

\item{mosquitos}{Names of mosquito species to simulate.}

\item{component}{Name of vector population intervention (i.e. LSM).}

\item{coverage}{A variable or a numeric value. Can be a placeholder.}

\item{decayVals}{Decay (if not specified, constant effectiveness, step
function as default)}

\item{startDate}{Date in YYYY-MM-DD format.}

\item{endDate}{Date in YYYY-MM-DD format.}

\item{interval}{A string like '1 weeks'. Same as in \code{\link[=seq.Date]{seq.Date()}}. Or a list
composed of the entries 'days' (optional), 'months' (optional) and
'years'. If a list is used, startDate and endDate are not used and can be
NULL.}

\item{dates}{If NULL, startDate, endDate and interval are used, else a
vector of dates in YYYY-MM-DD format. Can be a placeholder.}
}
\description{
Function that writes the larviciding component
}
