% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_interventions_define.R
\name{defineNothing}
\alias{defineNothing}
\alias{define_nothing}
\title{Writes an intervention parameterisation xml chunk that does nothing}
\usage{
defineNothing(baseList, mosquitos)

define_nothing(baseList, mosquitos)
}
\arguments{
\item{baseList}{List with experiment data.}

\item{mosquitos}{Name of mosquito species affected by the intervention.}
}
\description{
This is useful if something needs to be deployed, as a
placeholder.
}
