% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_interventions_define.R
\name{defineTreatSimple}
\alias{defineTreatSimple}
\alias{define_treatSimple}
\alias{define_treatSimple_compat}
\title{SMC, MDA parameterization section}
\usage{
defineTreatSimple(
  baseList,
  component = "MDA",
  durationBlood = "15d",
  durationLiver = 0
)

define_treatSimple(
  baseList,
  component = "MDA",
  durationBlood = "15d",
  durationLiver = 0
)

define_treatSimple_compat(
  baseList,
  component = "MDA",
  durationBlood = "15d",
  durationLiver = 0
)
}
\arguments{
\item{baseList}{List with experiment data.}

\item{component}{Name of the intervention, can be any name but needs to be
the same as defined in deployment}

\item{durationBlood}{Clearance of blood stage parasites}

\item{durationLiver}{Clearance of liver stage parasites}
}
\description{
Writes the treatSimple intervention used for mass treatments
(i.e. MDA, SMC)
}
