% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_interventions_define.R
\name{defineVaccine}
\alias{defineVaccine}
\alias{define_vaccine}
\title{Adds vaccine intervention parameterisation to baseList}
\usage{
defineVaccine(
  baseList,
  vaccineParameterization,
  append = TRUE,
  name = NULL,
  verbatim = FALSE,
  hist = FALSE
)

define_vaccine(
  baseList,
  vaccineParameterization,
  append = TRUE,
  name = NULL,
  verbatim = FALSE,
  hist = FALSE
)
}
\arguments{
\item{baseList}{List with experiment data.}

\item{vaccineParameterization}{Parameterization list}

\item{append}{If TRUE, then append to existing baseList, otherwise
overwrites}

\item{name}{Name tag list}

\item{verbatim}{If TRUE, then show messages}

\item{hist}{If TRUE, then decay is assumed to be step function set to 1 for
a year and then to zero for the remainder}
}
\description{
Adds vaccine intervention parameterisation to baseList
}
