% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_interventions_define.R
\name{defineVectorControl}
\alias{defineVectorControl}
\alias{define_vector_control}
\title{Adds vector control intervention parameterisation to baseList}
\usage{
defineVectorControl(
  baseList,
  vectorInterventionParameters,
  append = TRUE,
  name = NULL,
  verbatim = TRUE,
  hist = FALSE,
  resistance = 0.1
)

define_vector_control(
  baseList,
  vectorInterventionParameters,
  append = TRUE,
  name = NULL,
  verbatim = TRUE,
  hist = FALSE,
  resistance = 0.1
)
}
\arguments{
\item{baseList}{List with experiment data.}

\item{vectorInterventionParameters}{Vector control intervention
parameterization list depending on three parameters (deterrency,
preprandrial, postprandial) and decay functions:}

\item{append}{If TRUE, then append to existing baseList, otherwise
overwrites}

\item{name}{Name tag list}

\item{verbatim}{If TRUE, then show messages}

\item{hist}{If TRUE, then decay is assumed to be step function set to 1 for
a year and then to zero for the remainder}

\item{resistance}{Scaling function of insecticide resistance TODO}
}
\description{
Adds vector control intervention parameterisation to baseList
}
\examples{
vectorInterventionParameters <- list(
  "LLIN" = list(
    deterrency = list(
      decay = list(
        L = "0.7",
        "function" = "weibull"
      ),
      anophelesParams = list(
        "Anopheles gambiae" = list(
          propActive = 1,
          value = "0.5"
        ),
        "Anopheles funestus" = list(
          propActive = 1,
          value = "0.3"
        )
      )
    ),
    preprandialKillingEffect = list(
      decay = list(
        L = "0.4",
        "function" = "weibull"
      ),
      anophelesParams = list(
        "Anopheles gambiae" = list(
          propActive = 1,
          value = "0.6"
        ),
        "Anopheles funestus" = list(
          propActive = 1,
          value = "0.67"
        )
      )
    ),
    postprandialKillingEffect = list(
      decay = list(
        L = "0.45",
        "function" = "weibull"
      ),
      anophelesParams = list(
        "Anopheles gambiae" = list(
          propActive = 1,
          value = "0.3"
        ),
        "Anopheles funestus" = list(
          propActive = 1,
          value = "0.2"
        )
      )
    )
  )
)
}
