% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_interventions_deploy.R
\name{deploy_cont_compat}
\alias{deploy_cont_compat}
\title{Deployment of an continuous intervention}
\usage{
deploy_cont_compat(
  baseList,
  component = "IPTi",
  begin = "2019-01-01",
  end = "2030-01-01",
  targetAgeYrs = NULL,
  vaccMinPrevDoses = NULL,
  vaccMaxCumDoses = NULL,
  coverage = NULL,
  varyCov = FALSE,
  restrictToSubPop = NULL,
  cohort = NULL
)
}
\arguments{
\item{baseList}{List with experiment data.}

\item{component}{Name of intervention}

\item{begin}{Start date of the intervention (yyyy-mm-dd)}

\item{end}{End date of the intervention (yyyy-mm-dd)}

\item{targetAgeYrs}{Age at which intervention is received (1=1 year) e,g.
c(0.25,0.33 ,0.75)}

\item{vaccMinPrevDoses}{Applies to vaccines only: vaccine doses are only
deployed by this deployment if the previous number of doses (for the
component deployed) is at least this number. Needs to be of the same
length as targetAgeYrs.}

\item{vaccMaxCumDoses}{Applies to vaccines only: vaccine doses are only
deployed by this deployment if the previous number of doses (for the
component deployed) is less than this number. Needs to be of the same
length as targetAgeYrs.}

\item{coverage}{Coverage, either fixed (integer) or as a variable
("@IPTcov@"), same order as for target age years}

\item{varyCov}{Default is FALSE , for varying coverage}

\item{restrictToSubPop}{f this element is specified, deployment is
restricted to some sub-population}

\item{cohort}{Either NULL or character vector of cohort component ids.
Recipients of this deployment will also be recruited into these cohorts.}
}
\description{
Deployment of an continuous intervention
}
