% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_interventions_deploy.R
\name{deploy_it_compat}
\alias{deploy_it_compat}
\title{Writes the deployment of an intervention. Compatibility version.}
\usage{
deploy_it_compat(
  baseList,
  component = "ITN",
  cumulative = FALSE,
  effects = NULL,
  y1 = 2000,
  y2 = NULL,
  m1 = 1,
  m2 = NULL,
  d1 = 1,
  d2 = NULL,
  every = 1,
  interval = "month",
  SIMSTART = "1918-01-01",
  minAge = NULL,
  maxAge = NULL,
  coverage = NULL,
  byyear = FALSE,
  deployvar = NULL,
  subpop = FALSE,
  cohort = NULL
)
}
\arguments{
\item{baseList}{List with experiment data.}

\item{component}{Name of intervention.}

\item{cumulative}{default is FALSE. Do not set to TRUE.}

\item{effects}{Either NULL or c("det","pre","post")}

\item{y1}{Year of the first date (surveys starting from year y1)}

\item{y2}{Year of the end date (surveys continuing until year y2)}

\item{m1}{Month of the first date}

\item{m2}{Month of the end date}

\item{d1}{Day of the first date}

\item{d2}{Day of the end date}

\item{every}{Interval size}

\item{interval}{Interval size (days, weeks, )}

\item{SIMSTART}{Starting date of the simulations in the format "yyyy-mm-dd"}

\item{minAge}{Minimum age for deployment (used in SMC)}

\item{maxAge}{Maximum age for deployment (used in SMC)}

\item{coverage}{Value or variable of coverage}

\item{byyear}{If TRUE, allows coverage to vary by year
('histITNcov2000',...)}

\item{deployvar}{Allows for deployment dates to vary (across years y1, ...,
y2)}

\item{subpop}{If TRUE, then restricts to a subpopulation (see
restrictToSubPop in OpenMalaria)}

\item{cohort}{Either NULL or character vector of cohort component ids.
Recipients of this deployment will also be recruited into these cohorts.}
}
\description{
Writes the deployment of an intervention. Compatibility version.
}
