% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_interventions_deploy.R
\name{deployIT}
\alias{deployIT}
\alias{deploy_IT}
\title{Writes the deployment of an intervention.}
\usage{
deployIT(
  baseList,
  component = "ITN",
  cumulative = FALSE,
  effects = NULL,
  startDate = NULL,
  endDate = NULL,
  interval = NULL,
  dates = NULL,
  minAge = NULL,
  maxAge = NULL,
  coverage = NULL,
  subpop = NULL,
  cohort = NULL
)

deploy_IT(
  baseList,
  component = "ITN",
  cumulative = FALSE,
  effects = NULL,
  startDate = NULL,
  endDate = NULL,
  interval = NULL,
  dates = NULL,
  minAge = NULL,
  maxAge = NULL,
  coverage = NULL,
  subpop = NULL,
  cohort = NULL
)
}
\arguments{
\item{baseList}{List with experiment data.}

\item{component}{Name of intervention.}

\item{cumulative}{Default is FALSE. Do not set to TRUE.}

\item{effects}{Either NULL or vector of strings, e.g. c("det","pre","post")}

\item{startDate}{Date in YYYY-MM-DD format.}

\item{endDate}{Date in YYYY-MM-DD format.}

\item{interval}{A string like '1 weeks'. Same as in \code{\link[=seq.Date]{seq.Date()}}. Or a list
composed of the entries 'days' (optional), 'months' (optional) and
'years'. If a list is used, startDate and endDate are not used and can be
NULL.}

\item{dates}{If NULL, startDate, endDate and interval are used, else a
vector of dates in YYYY-MM-DD format. Can be a placeholder.}

\item{minAge}{Minimum age for deployment (used in SMC). Can be a
placeholder.}

\item{maxAge}{Maximum age for deployment (used in SMC). Can be a
placeholder.}

\item{coverage}{Value or variable of coverage. Can be a placeholder.}

\item{subpop}{Either NULL or string. Concatenation component+"-"+subpop will
be id in subpopulation intervention should be restricted to (see
restrictToSubPop in OpenMalaria)}

\item{cohort}{Either NULL or character vector of cohort component ids.
Recipients of this deployment will also be recruited into these cohorts.}
}
\description{
Writes the deployment of an intervention.
}
