% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{do_post_process_cleanup}
\alias{do_post_process_cleanup}
\title{Cleaning up the parallel version of CombinedDat_Aggr,
CombinedDat_wide}
\usage{
do_post_process_cleanup(
  nameExperiment,
  widename = "_CombinedDat_wide.RData",
  aggname = "_CombinedDat_Aggr.RData",
  monthname = "_CombinedDat_month.RData",
  fut = "fut",
  removefiles = TRUE,
  make_aggr = TRUE,
  make_wide = TRUE,
  make_month = TRUE,
  setting = "setting",
  setting_number = 1,
  seed_as_hist_param = TRUE,
  placeholder = NULL,
  include = NULL,
  debugg = FALSE
)
}
\arguments{
\item{nameExperiment}{name of the experiment (string)}

\item{widename}{Name convention for CombinedDat_wide files}

\item{aggname}{Name convention for CombinedDat_Aggr files}

\item{monthname}{Name of dataframe (i.e. "_CombinedDat_month.Rdata")}

\item{fut}{Name of future prefix, default is 'fut'}

\item{removefiles}{Removes the temporary files}

\item{make_aggr}{Combines the CombinedDat_Aggr files}

\item{make_wide}{If T, creates a wide dataset}

\item{make_month}{If TRUE, writes CombinedDat_month files}

\item{setting}{Variable name that specifies the setting to run (default
'setting')}

\item{setting_number}{Loop index for setting 1:N}

\item{seed_as_hist_param}{If TRUE, each seed has a different past value}

\item{placeholder}{Variables that are neither historical nor future nor unique scenario identifiers}

\item{include}{Variables that should be unique scenario identifiers, but are not in 'full'}

\item{debugg}{If TRUE, then runs only 30 files from each setting}
}
\description{
Cleaning up the parallel version of CombinedDat_Aggr,
CombinedDat_wide
}
\note{
used often internally, never seen by user, easy to write, rarely
breaks
}
