% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{do_post_processing}
\alias{do_post_processing}
\title{Function to manage postprocessing The output file is named
paste0(setting_number,"_",loop_id,"_CombinedDat_Aggr.RData") and is generated
for each scenario and setting separately}
\usage{
do_post_processing(
  nameExperiment,
  basename,
  ORIGIN = "1918-01-01",
  number_loops = 12,
  setting_number = 1,
  loop_id = 1,
  fsize = 500,
  timeunit = "year",
  fut = "fut",
  aggregate_to_year = TRUE,
  agecats = c("0to5", "6to11", "2to10", "All"),
  ignores = c("propOut", "jan", "feb", "mar", "apr", "may", "jun", "jul", "aug", "sep",
    "oct", "nov", "dec", "scaled_down_flag", paste0("m", 1:12), paste0("month", 1:12)),
  selectedOutVars = c("PR", "nUncomp", "nHost", "nSevere", "incidence", "ddeath",
    "edeath", "tUncomp", "tSevere", "nHosp"),
  debugg = FALSE,
  widename = "_CombinedDat_wide.RData",
  setting = "setting",
  seed_as_hist_param = TRUE,
  monthvars = c("PR", "incidence"),
  monthname = "_CombinedDat_month.RData",
  monthyears = 2015:2030,
  placeholder = NULL,
  include = NULL
)
}
\arguments{
\item{nameExperiment}{name of the experiment (string)}

\item{basename}{name of the base xml file (i.e. "base.xml")}

\item{ORIGIN}{date when simulation starts (yyyy-mm-dd), default =
"1918-01-01"}

\item{number_loops}{defines how many loops to run in parallel (i.e. 12)}

\item{setting_number}{which loop is this one for?}

\item{loop_id}{which id number (sets of 500 files )}

\item{fsize}{how many files to combine}

\item{timeunit}{"year" is default}

\item{fut}{future variables start with 'fut'}

\item{aggregate_to_year}{if TRUE, it aggregates months to year}

\item{agecats}{defines for which age groups to postprocess, options include:
c("0to5","6to11","2to10","All")}

\item{ignores}{variables to ignore (e.g. m1, m2, m3, m4, )}

\item{selectedOutVars}{variables of interest}

\item{debugg}{if TRUE, function runs in debugging mode (very small dataset)}

\item{widename}{name of dataframe (i.e. "_CombinedDat_wide.RData")}

\item{setting}{variable name that specifies the setting (default: "setting")}

\item{seed_as_hist_param}{TODO}

\item{monthvars}{TODO}

\item{monthname}{name of dataframe (i.e. "_CombinedDat_month.RData")}

\item{monthyears}{which years to keep in the year.month (very wide) dataset}

\item{placeholder}{TODO}

\item{include}{TODO}
}
\description{
Function to manage postprocessing The output file is named
paste0(setting_number,"_",loop_id,"_CombinedDat_Aggr.RData") and is generated
for each scenario and setting separately
}
\note{
used often internally, never seen by user, difficult to write,
sometimes breaks
}
