% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{.add_file_column_to_scens}
\alias{.add_file_column_to_scens}
\title{Helper function for write_scen_data}
\usage{
.add_file_column_to_scens(scens, nameExperiment, startnum = 1)
}
\arguments{
\item{scens}{scenario object}

\item{nameExperiment}{name of experiment}

\item{startnum}{starting number of scenario files}
}
\description{
Helper function for write_scen_data
}
