% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{.addExpToDB}
\alias{.addExpToDB}
\title{Add data to experiments table in DB}
\usage{
.addExpToDB(connection, x, method = "ignore")
}
\arguments{
\item{connection}{Database connection.}

\item{x}{Data to add.}

\item{method}{How to handle duplicate date. Can be "ignore" or "replace".}
}
\description{
Add data to experiments table in DB
}
\keyword{internal}
