% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{.addPlaceholdersToDB}
\alias{.addPlaceholdersToDB}
\title{Add data to placeholders table in DB}
\usage{
.addPlaceholdersToDB(connection, x)
}
\arguments{
\item{connection}{Database connection.}

\item{x}{Data to add.}
}
\description{
Add data to placeholders table in DB
}
\keyword{internal}
