% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{.addResultsToDB}
\alias{.addResultsToDB}
\title{Add data to results table in DB}
\usage{
.addResultsToDB(connection, x)
}
\arguments{
\item{connection}{Database connection.}

\item{x}{Data to add.}
}
\description{
Add data to results table in DB
}
\keyword{internal}
