% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{.addScenToDB}
\alias{.addScenToDB}
\title{Add data to scenarios and scenarios_metadata table in DB}
\usage{
.addScenToDB(connection, x)
}
\arguments{
\item{connection}{Database connection.}

\item{x}{Data to add.}
}
\description{
Add data to scenarios and scenarios_metadata table in DB
}
\keyword{internal}
