% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess-compat.R
\name{.aggregate_age_groups}
\alias{.aggregate_age_groups}
\title{generate age sets..}
\usage{
.aggregate_age_groups(
  ages,
  rawdat,
  groups = c("0to5", "6to11", "2to10", "All"),
  group_variables,
  outcome_variables
)
}
\arguments{
\item{ages}{vector of age labels}

\item{rawdat}{rawdat object}

\item{groups}{default groups used  "0to5","2to10","6to11","All"}

\item{group_variables}{example could be "age" "year" "setting"}

\item{outcome_variables}{"PR" or "incidence"}
}
\description{
generate age sets..
}
\note{
used often internally, never seen by user, easy to write, never breaks
}
