% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess-compat.R
\name{.aggregate_to_year}
\alias{.aggregate_to_year}
\title{aggregate survey to year - (depending on frequency of surveys timesteps monitored)}
\usage{
.aggregate_to_year(
  CombinedDat,
  age_variables = "age",
  time_variables = "year",
  outcome_variables,
  model_variables,
  setting_variables
)
}
\arguments{
\item{CombinedDat}{dataset}

\item{age_variables}{"age"}

\item{time_variables}{"year"}

\item{outcome_variables}{"PR"}

\item{model_variables}{model_variables = c("models", "seed", "scenario")}

\item{setting_variables}{"setting"}
}
\description{
aggregate survey to year - (depending on frequency of surveys timesteps monitored)
}
\note{
used often internally, never seen by user, medium to write, never breaks
}
