% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{.assign_id_variables}
\alias{.assign_id_variables}
\title{Function to generate and add identifier variables for history, future and
all scenarios}
\usage{
.assign_id_variables(
  CombinedDat_wide = NULL,
  unique_variables = NULL,
  historical_variables = NULL,
  future_variables = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{CombinedDat_wide}{CombinedDat_wide dataset}

\item{unique_variables}{Values of unique variables}

\item{historical_variables}{Values of historical variables}

\item{future_variables}{Values of future variables}

\item{overwrite}{Default to FALSE, if want to overwrite existing numbers}
}
\description{
Function to generate and add identifier variables for history, future and
all scenarios
}
\note{
Used often internally, never seen by user, easy to write, sometimes
breaks
}
