% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm_postprocessing.R
\name{.assign_post_process_loop}
\alias{.assign_post_process_loop}
\title{Assigns the number of loops needed}
\usage{
.assign_post_process_loop(scens, fsize, setting = "setting")
}
\arguments{
\item{scens}{scens object}

\item{fsize}{number of files to merge}

\item{setting}{"setting"}
}
\description{
Assigns the number of loops needed
}
\keyword{internal}
