% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm_postprocessing.R
\name{.calculate_postprocess_loops}
\alias{.calculate_postprocess_loops}
\title{helper function for batch_process}
\usage{
.calculate_postprocess_loops(scens, fsize, setting = "setting")
}
\arguments{
\item{scens}{scens object}

\item{fsize}{number of files to merge}

\item{setting}{"setting"}
}
\description{
helper function for batch_process
}
\keyword{internal}
