% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_interventions_define.R
\name{.calculateSmoothCompact}
\alias{.calculateSmoothCompact}
\title{Calculate smooth compact decay parameters}
\usage{
.calculateSmoothCompact(k, L, t = seq(0, 10, length.out = 100))
}
\arguments{
\item{k}{Shape parameter}

\item{L}{Decay parameter (number of years)}

\item{t}{Vector of time (years)}
}
\description{
Calculate smooth compact decay parameters
}
