% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_interventions_define.R
\name{.calculateSmoothCompactHalflife}
\alias{.calculateSmoothCompactHalflife}
\title{Calculate smooth compact halflife}
\usage{
.calculateSmoothCompactHalflife(
  k = 2.14,
  L = 6.08,
  halflife = 2,
  threshold = 0.5,
  grid_n = 50
)
}
\arguments{
\item{k}{Scaling factor in smooth compact}

\item{L}{Other factor in smooth compact}

\item{halflife}{Desired halflife (positive number)}

\item{threshold}{Set to 0.50 for 'halflife'}

\item{grid_n}{How fine of a grid to explore (50 x 50)}
}
\description{
Explores a space of k x L to find the smooth-compact closest to
the given smooth-compact, but with a halflife of 'halflife' (say 2 instead
of 3)
}
