% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.compressFiles}
\alias{.compressFiles}
\title{Compress files into zip archive}
\usage{
.compressFiles(dir, pattern = NULL, remove = FALSE)
}
\arguments{
\item{dir}{Path to directory.}

\item{pattern}{An optional regular expression. See \link[base]{list.files}}

\item{value}{If TRUE, files will be removed after adding them to the
archive.}
}
\description{
Compress files into zip archive
}
\keyword{internal}
