% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{.createDB}
\alias{.createDB}
\title{Create a database connection}
\usage{
.createDB(dbName, path = getCache("rootDir"))
}
\arguments{
\item{dbName}{Name of the database file without extension.}

\item{path}{Directory of the database file. Defaults to the root directory.}
}
\description{
Create a SQLite database in the root directory if it does not
exist and return the connection.
}
\keyword{internal}
