% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{.createResultsTable}
\alias{.createResultsTable}
\title{Create the DB result table}
\usage{
.createResultsTable(connection, tName, columns)
}
\arguments{
\item{connection}{Database connection.}

\item{tName}{Name of the results table.}

\item{columns}{A list containing the column names and types, e.g. list(names
= c(scenario_id, value), types = c("INTEGER", "NUMERIC")). The column
"experiment_id" is always automatically added.}
}
\description{
Creates the database schema for the results.
}
\keyword{internal}
