% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess-compat.R
\name{.define_new_outcome_variable}
\alias{.define_new_outcome_variable}
\title{Function to calculate new outcome variables}
\usage{
.define_new_outcome_variable(
  dat,
  newvar,
  sumvar,
  divideby = NULL,
  units_of = 1
)
}
\arguments{
\item{dat}{rawdat}

\item{newvar}{variable name to create}

\item{sumvar}{variables to sum over}

\item{divideby}{divide by (column name)}

\item{units_of}{report in units of units_of (1000)}
}
\description{
Function to calculate new outcome variables
}
\note{
used often internally, never seen by user, easy to write, never breaks
}
