% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_interventions_deploy.R
\name{.equalizePlaceholders}
\alias{.equalizePlaceholders}
\title{Equalize length of placeholder sequences}
\usage{
.equalizePlaceholders(x, placeholderseq)
}
\arguments{
\item{x}{Sequence which should be the maximum lenght.}

\item{placeholderseq}{List with placeholder sequences.}
}
\description{
Equalize length of placeholder sequences
}
\keyword{internal}
