% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess-compat.R
\name{.extract_base_param}
\alias{.extract_base_param}
\title{Supplementary r-script to be included by others, to load key information from
base xml}
\usage{
.extract_base_param(baseXml = NULL, SIMSTART = "1918-01-01", pop = NULL)
}
\arguments{
\item{baseXml}{Name of the base xml file}

\item{SIMSTART}{Starting date of the simulations in the format "yyyy-mm-dd",
default e.g. "1918-01-01", If NULL the base xml is searched for the term
"Simulation Start" or surveys number 1}

\item{pop}{Total population (integer), if NULL, it will be read from the xml
file}
}
\description{
Supplementary r-script to be included by others, to load key information from
base xml
}
\note{
used often internally, never seen by user, easy to write, never breaks
}
