% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess-compat.R
\name{.get_om_event_measure}
\alias{.get_om_event_measure}
\title{Function to get event measures names and corresponding numbers ' @note used}
\usage{
.get_om_event_measure()
}
\description{
Function to get event measures names and corresponding numbers ' @note used
}
