% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getIndex}
\alias{.getIndex}
\title{Search for name and/or value}
\usage{
.getIndex(l, name = NULL, value = NULL)
}
\arguments{
\item{l}{Input list.}

\item{name}{Name to look for.}

\item{value}{Value to look for.}
}
\description{
Search for name and/or value
}
\keyword{internal}
