% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{.identify_files_to_join}
\alias{.identify_files_to_join}
\title{Helper function in do_post_process_cleanup}
\usage{
.identify_files_to_join(
  CombinedDir,
  setting_number = 1,
  widename = "_CombinedDat_wide.RData"
)
}
\arguments{
\item{CombinedDir}{folder with CombinedDat_wide files in it}

\item{setting_number}{setting number (1, 2, 3, ..., 11)}

\item{widename}{"CombinedDat_wide.RData"}
}
\description{
Helper function in do_post_process_cleanup
}
\note{
internal function
}
