% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.indexList}
\alias{.indexList}
\title{Build index of a list recursively}
\usage{
.indexList(l, col = list(), n = NULL)
}
\arguments{
\item{l}{Input list.}

\item{col}{Collection list.}

\item{n}{Current index level.}
}
\description{
Build index of a list recursively
}
\keyword{internal}
