% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess-compat.R
\name{.loop_over_output_files}
\alias{.loop_over_output_files}
\title{returns a list of files}
\usage{
.loop_over_output_files(
  scens,
  setting = "setting",
  files,
  setting_number,
  number_loops
)
}
\arguments{
\item{scens}{scens object}

\item{setting}{variable name that specifies the setting to run (default 'setting')}

\item{files}{list of files to loop over}

\item{setting_number}{which department/Setting number}

\item{number_loops}{how many loops in total}
}
\description{
returns a list of files
}
