% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_root.R
\name{.makeXmlRoot}
\alias{.makeXmlRoot}
\title{Generate XML root}
\usage{
.makeXmlRoot(schemaVersion, name, analysisNo)
}
\arguments{
\item{schemaVersion}{OM schema version}

\item{name}{Experiment name}

\item{analysisNo}{Analysis number}
}
\description{
Generate XML root
}
\keyword{internal}
