% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess-new.R
\name{.measuresNeededForPostprocessing}
\alias{.measuresNeededForPostprocessing}
\title{List of measures needed to calculate epidemiological indicators}
\usage{
.measuresNeededForPostprocessing()
}
\description{
List of measures needed to calculate epidemiological indicators
}
\keyword{internal}
