% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess-compat.R
\name{.merge_scens_with_outputs}
\alias{.merge_scens_with_outputs}
\title{merging walle with scenarios columns
merge "files_out" list (in that order) with the variable values
specified in the scenarios.csv file}
\usage{
.merge_scens_with_outputs(
  walle,
  short_filename,
  scens,
  ignore = c("propOut", "jan", "feb", "mar", "apr", "may", "jun", "jul", "aug", "sep",
    "oct", "nov", "dec", "scaled_down_flag")
)
}
\arguments{
\item{walle}{wide object}

\item{short_filename}{"wutest_1" "wutest_2", short in that it does not have ".xml" at end}

\item{scens}{dataset of scenarios}

\item{ignore}{variables to ignore}
}
\description{
merging walle with scenarios columns
merge "files_out" list (in that order) with the variable values
specified in the scenarios.csv file
}
\note{
used often internally, never seen by user, easy to write, never breaks
}
