% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess-compat.R
\name{.merge_with_age_time}
\alias{.merge_with_age_time}
\title{add age group information to dataset}
\usage{
.merge_with_age_time(rawdat, age_dataframe, timestep_dataframe, nHost)
}
\arguments{
\item{rawdat}{dataset}

\item{age_dataframe}{created from .extract_base_param()}

\item{timestep_dataframe}{created from .extract_base_param()}

\item{nHost}{variable of host}
}
\description{
add age group information to dataset
}
\note{
used often internally, never seen by user, easy to write, never breaks
}
