% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{.numberToSurveyMeasure}
\alias{.numberToSurveyMeasure}
\title{Dictionary mapping survey measure numbers to names}
\usage{
.numberToSurveyMeasure()
}
\description{
See:
https://github.com/SwissTPH/openmalaria/wiki/MonitoringOptions
}
\keyword{internal}
