% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_interventions_deploy.R
\name{.placeholderseqGen}
\alias{.placeholderseqGen}
\title{Generate placeholder sequence}
\usage{
.placeholderseqGen(x, placeholders)
}
\arguments{
\item{x}{Input data. It is assumed that this is a list is in the form of
'list(foo = list("bar", c(1, 2, 3, ...)), ...)' or 'list(foo = c(seq1,
seq2, ...), ...)'}

\item{placeholders}{Character vector of names which should be processed}
}
\description{
Generate placeholder sequence
}
\keyword{internal}
