% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing.R
\name{.printDebug}
\alias{.printDebug}
\title{Print debug messages}
\usage{
.printDebug(x, toggle = NULL, ...)
}
\arguments{
\item{x}{Object to print.}

\item{toggle}{TRUE, FALSE or a boolean variable. If NULL, the state of the
package variable 'verboseOutput' is used.}

\item{...}{Further arguments to \code{\link[=print]{print()}}.}
}
\description{
Like print, but only if 'debugOutput' is TRUE.
}
\keyword{internal}
