% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.readCache}
\alias{.readCache}
\title{Read cache from disk}
\usage{
.readCache(path)
}
\arguments{
\item{path}{Path to experiment.}
}
\description{
Read cache from 'path'. The folder structure 'cache/cache.rds'
is appended.
}
\keyword{internal}
