% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{.scenariosFilenames}
\alias{.scenariosFilenames}
\title{Add a filename column to scenarios}
\usage{
.scenariosFilenames(scenarios, prefix)
}
\arguments{
\item{scenarios}{Data frame containing the values for the placeholders. One
row per scenario, placeholders in columns. Column names correspond to the
placeholder names.}

\item{prefix}{Filename prefix}
}
\description{
Add a filename column to scenarios
}
\keyword{internal}
