% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{.scenariosRowSelect}
\alias{.scenariosRowSelect}
\title{Select rows for scenario creation}
\usage{
.scenariosRowSelect(scenarios, rowStart = NULL, rowEnd = NULL)
}
\arguments{
\item{scenarios}{Data frame containing the values for the placeholders. One
row per scenario, placeholders in columns. Column names correspond to the
placeholder names.}

\item{rowStart}{Starting row. Optional.}

\item{rowEnd}{End row. Optional.}
}
\description{
Select rows for scenario creation
}
\keyword{internal}
