% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{.seed_warning}
\alias{.seed_warning}
\title{Internal warning funtion}
\usage{
.seed_warning(scens, seed_as_hist_param)
}
\arguments{
\item{scens}{scens object}

\item{seed_as_hist_param}{if TRUE, each seed has a different past value}
}
\description{
Internal warning funtion
}
