% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{.setFunArgs}
\alias{.setFunArgs}
\title{Override default function arguments}
\usage{
.setFunArgs(f, args)
}
\arguments{
\item{f}{Function}

\item{args}{List of function arguments}
}
\description{
Override default function arguments
}
\keyword{internal}
