% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess-compat.R
\name{.shorten_output_files}
\alias{.shorten_output_files}
\title{returns 'alle' with columns := survey, age_group,measure,value,scenario}
\usage{
.shorten_output_files(
  subset_of_files,
  fsize = 100,
  sets = T,
  one_setting,
  setting_number = 1,
  MalariaDir,
  loop_id = 1
)
}
\arguments{
\item{subset_of_files}{list of files}

\item{fsize}{maximum number of files to combine}

\item{sets}{T or F if scens has setting column}

\item{one_setting}{name of setting variable}

\item{setting_number}{Setting number index}

\item{MalariaDir}{directory of output files}

\item{loop_id}{loop_id for each 500 files}
}
\description{
returns 'alle' with columns := survey, age_group,measure,value,scenario
}
\note{
used often internally, never seen by user, medium to write, never breaks
}
