% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm.R
\name{.slurmOptions}
\alias{.slurmOptions}
\title{Generate top section of SLURM script}
\usage{
.slurmOptions(
  jobName,
  ntasks = NULL,
  dependency = NULL,
  memCPU = NULL,
  output = NULL,
  error = NULL,
  array = NULL,
  time = NULL,
  qos = NULL
)
}
\arguments{
\item{jobName}{Name of the job}

\item{ntasks}{Number of tasks}

\item{dependency}{Job ids completed after which execution should start}

\item{memCPU}{RAM per CPU}

\item{output}{Output file}

\item{error}{Error log file}

\item{array}{Array range}

\item{time}{Maximum time}

\item{qos}{Quality of service}
}
\description{
Generate top section of SLURM script
}
\keyword{internal}
