% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess-compat.R
\name{.spread_across_columns}
\alias{.spread_across_columns}
\title{Function to spread for multiple columns
Obtained from the r studio community: https://community.rstudio.com/t/spread-with-multiple-value-columns/5378}
\usage{
.spread_across_columns(df, key, value)
}
\arguments{
\item{df}{dataframe}

\item{key}{column with variable}

\item{value}{column of value}
}
\description{
Function to spread for multiple columns
Obtained from the r studio community: https://community.rstudio.com/t/spread-with-multiple-value-columns/5378
}
\note{
used often internally, never seen by user, easy to write, never breaks
}
