% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess-compat.R
\name{.spreadAcrossColumns}
\alias{.spreadAcrossColumns}
\title{Function to spread for multiple columns
Based on: https://community.rstudio.com/t/spread-with-multiple-value-columns/5378}
\usage{
.spreadAcrossColumns(df, key, value)
}
\arguments{
\item{df}{Dataframe}

\item{key}{Variable column}

\item{value}{Value columns}
}
\description{
Function to spread for multiple columns
Based on: https://community.rstudio.com/t/spread-with-multiple-value-columns/5378
}
