% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess-compat.R
\name{.sum_over_age_groups}
\alias{.sum_over_age_groups}
\title{Collate results per  age group}
\usage{
.sum_over_age_groups(
  rawdat,
  age_range,
  age_group_character,
  group_variables,
  outcome_variables
)
}
\arguments{
\item{rawdat}{rawdata dataset}

\item{age_range}{"0-1" as example, dash is important}

\item{age_group_character}{"0to1" "All" as examples}

\item{group_variables}{group variables (year, setting, etc)}

\item{outcome_variables}{variables of interest}
}
\description{
Collate results per  age group
}
\note{
used often internally, never seen by user, easy to write, never breaks
}
