% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.syncBoth}
\alias{.syncBoth}
\title{Synchronize diskCache and memCache}
\usage{
.syncBoth(diskCache, memCache)
}
\arguments{
\item{diskCache}{Cache environment}

\item{memCache}{Cache environment}
}
\description{
Synchronize diskCache and memCache
}
\keyword{internal}
