% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.synchronizeCache}
\alias{.synchronizeCache}
\title{Synchronize the Cache}
\usage{
.synchronizeCache(path = NULL, direction = c("disk", "memory", "none"))
}
\arguments{
\item{path}{Path to cache directory.}

\item{direction}{Directions of sync. Can be one of "disk" = memory cache
overrides disk cache, "memory" = disk cache overrides memory cache and
"none" = caches get synchronized via timestamps.}
}
\description{
Synchronize the Cache
}
\keyword{internal}
