% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.syncMem}
\alias{.syncMem}
\title{Assign values from diskCache to .pkgcache}
\usage{
.syncMem(diskCache)
}
\arguments{
\item{diskCache}{Cache environment}
}
\description{
Assign values from diskCache to .pkgcache
}
\keyword{internal}
