% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess-compat.R
\name{.transform_wide_to_long}
\alias{.transform_wide_to_long}
\title{Transform dataset from wide to long format}
\usage{
.transform_wide_to_long(
  dataframe,
  timevar,
  outcomes_of_interest = outcome_variables,
  model_variables,
  setting_variables
)
}
\arguments{
\item{dataframe}{default CombinedDat_Aggr}

\item{timevar}{default timeunit}

\item{outcomes_of_interest}{default  selectedOutVars}

\item{model_variables}{model_variables = c("models", "seed", "scenario")}

\item{setting_variables}{example is "setting"}
}
\description{
Transform dataset from wide to long format
}
\note{
used often internally, never seen by user, medium to write, never breaks
}
