% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess-compat.R
\name{.two_digit_month}
\alias{.two_digit_month}
\title{uses the month of a date to create a two digit month (character)}
\usage{
.two_digit_month(CombinedDat)
}
\arguments{
\item{CombinedDat}{CombinedDat dataset}
}
\description{
uses the month of a date to create a two digit month (character)
}
\note{
used internally, never breaks, never seen by user
}
