% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess-compat.R
\name{.twoDigitMonth}
\alias{.twoDigitMonth}
\title{Uses the month of a date to create a two digit month (character)}
\usage{
.twoDigitMonth(CombinedDat)
}
\arguments{
\item{CombinedDat}{CombinedDat dataset}
}
\description{
Uses the month of a date to create a two digit month (character)
}
