% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directories.R
\name{.useDir}
\alias{.useDir}
\title{Create a folder and return the path}
\usage{
.useDir(path)
}
\arguments{
\item{path}{Path of directory.}
}
\description{
If a folder does not exist, create it and return the path.
Otherwise, just return the path.
}
\keyword{internal}
