% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{.warn_about_bad_names}
\alias{.warn_about_bad_names}
\title{Warns about bad choice of names in CSV files}
\usage{
.warn_about_bad_names(dat, setting = "setting")
}
\arguments{
\item{dat}{A csv file that will be imported.}

\item{setting}{"setting" as default.}
}
\description{
Warns about bad choice of names in CSV files
}
