% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess-compat.R
\name{.widen_processed_dataset}
\alias{.widen_processed_dataset}
\title{creates a wide dataset in post-processing}
\usage{
.widen_processed_dataset(alle, match_measure_to_number)
}
\arguments{
\item{alle}{dataset}

\item{match_measure_to_number}{output obtained from .get_om_event_measure()}
}
\description{
creates a wide dataset in post-processing
}
\note{
used often internally, never seen by user, easy to write, never breaks
}
