% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{.widenProcessedDataset}
\alias{.widenProcessedDataset}
\title{Creates a wide dataset in post-processing}
\usage{
.widenProcessedDataset(data, matchMeasureToNumber)
}
\arguments{
\item{data}{Dataset}

\item{matchMeasureToNumber}{Output obtained from .surveyMeasuresDict()}
}
\description{
Creates a wide dataset in post-processing
}
