% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm.R
\name{.writeSlurm}
\alias{.writeSlurm}
\title{Write SLURM script to file}
\usage{
.writeSlurm(
  jobName,
  ntasks = NULL,
  memCPU = NULL,
  output = NULL,
  error = NULL,
  array = NULL,
  time = NULL,
  qos = NULL,
  pre = NULL,
  cmd = NULL,
  post = NULL,
  file = NULL
)
}
\arguments{
\item{jobName}{Name of the job}

\item{ntasks}{Number of tasks}

\item{memCPU}{RAM per CPU}

\item{output}{Output file}

\item{error}{Error log file}

\item{array}{Array range}

\item{time}{Maximum time}

\item{qos}{Quality of service}

\item{pre}{List of commands before main command}

\item{cmd}{Main command}

\item{post}{List of commands after main command}

\item{file}{Filename}
}
\description{
Write SLURM script to file
}
\keyword{internal}
