% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{.xmlAddChunks}
\alias{.xmlAddChunks}
\title{Generate XML chunks}
\usage{
.xmlAddChunks(outlist, element, attributeList)
}
\arguments{
\item{outlist}{TODO}

\item{element}{TODO}

\item{attributeList}{TODO}
}
\value{
XML entries
}
\description{
Generate XML chunks
}
