% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{.xmlMakeDocRec}
\alias{.xmlMakeDocRec}
\title{Create xml entries recursively}
\usage{
.xmlMakeDocRec(baseXML, data, validate = TRUE)
}
\arguments{
\item{baseXML}{The root xml object}

\item{data}{List with experiment information}

\item{validate}{Boolean, if data should be validated}
}
\value{
xml document
}
\description{
Create xml entries recursively
}
