% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_monitoring.R
\name{.xmlMonitoringTimeRegularSeq}
\alias{.xmlMonitoringTimeRegularSeq}
\title{Generate a date sequence}
\usage{
.xmlMonitoringTimeRegularSeq(startDate, endDate, daysFilter)
}
\arguments{
\item{startDate}{Start date in 'YYYY-MM-DD'}

\item{endDate}{End date in 'YYYY-MM-DD'}

\item{daysFilter}{Day differences to keep}
}
\description{
Generate a date sequence
}
\keyword{internal}
