% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{.xmlTimeBlockSeq}
\alias{.xmlTimeBlockSeq}
\title{Generate date sequence based on specified days, months and years}
\usage{
.xmlTimeBlockSeq(interval)
}
\arguments{
\item{interval}{A list composed of the entries 'days' (optional), 'months'
(optional) and 'years'}
}
\value{
A sequence of dates in the YYYY-MM-DD format
}
\description{
Generate date sequence based on specified days, months and years
}
