% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{.xmlTimeRegularSeq}
\alias{.xmlTimeRegularSeq}
\title{Generate date sequence based on start and end date}
\usage{
.xmlTimeRegularSeq(startDate, endDate, interval)
}
\arguments{
\item{startDate}{Date in YYYY-MM-DD format}

\item{endDate}{Date in YYYY-MM-DD format}

\item{interval}{A string like '1 weeks'. Same as in \code{\link[=seq.Date]{seq.Date()}}}
}
\value{
A sequence of dates in the YYYY-MM-DD format
}
\description{
Generate date sequence based on start and end date
}
