% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{enclose}
\alias{enclose}
\title{Enclose a string of text and passing it on as a vector c("blah")}
\usage{
enclose(text)
}
\arguments{
\item{text}{Text to enclose}
}
\description{
Enclose a string of text and passing it on as a vector c("blah")
}
\examples{
stuff <- c("a", "b", "c", "d")
enclose(stuff)
}
