% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{enclose_numeric}
\alias{enclose_numeric}
\title{Enclose (numeric)}
\usage{
enclose_numeric(num)
}
\arguments{
\item{num}{numeric values}
}
\description{
To enclose a string of num and pass it on as a c(1,2,3) form
}
