% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expInfo}
\alias{expInfo}
\title{Collect information about current experiment}
\usage{
expInfo(what = NULL, print = TRUE)
}
\arguments{
\item{what}{Which information to collect. Can be any of c("name", "path",
"om_version", "placeholders", "simStart", "surveyTimes")}

\item{print}{If TRUE, pretty print the information. Otherwise return a list.}
}
\description{
Collect information about current experiment
}
