% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extractList}
\alias{extractList}
\title{Extract sublist which matches input}
\usage{
extractList(l, name = NULL, value = NULL, onlyIndex = FALSE)
}
\arguments{
\item{l}{Input list.}

\item{name}{Name to look for.}

\item{value}{Value to look for.}

\item{onlyIndex}{If TRUE, only return numeric index of matches.}
}
\description{
Extract sublist which matches input
}
