% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{finalizeScenarios}
\alias{finalizeScenarios}
\alias{finalize_scenarios}
\title{Generate scenarios from a base xml file}
\usage{
finalizeScenarios(x)

finalize_scenarios(x)
}
\arguments{
\item{x}{Data frame containing the values for the placeholders for each
scenarios.}
}
\description{
Function makes sure that the scenario data frame is set up
correctly. It adds the required 'ID' and 'file' columns.
}
