% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{generateScenarios}
\alias{generateScenarios}
\alias{generate_scenarios}
\title{Generate scenarios from a base xml file}
\usage{
generateScenarios(x)

generate_scenarios(x)
}
\arguments{
\item{x}{Data frame containing the values for the placeholders for each
scenarios.}
}
\description{
Function makes sure that the scenario data frame is set up
correctly. It adds the required 'ID' and 'file' columns.
}
