% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GHA}
\alias{GHA}
\title{Demographic data from Ghana}
\format{
An object of class \code{data.frame} with 20 rows and 2 columns.
}
\usage{
GHA
}
\description{
A dataset with demographic data from Ghana. The variables are as follows:
}
\keyword{datasets}
