% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HTI}
\alias{HTI}
\title{Demographic data from Haiti}
\format{
An object of class \code{data.frame} with 18 rows and 2 columns.
}
\usage{
HTI
}
\description{
A dataset with demographic data from Haiti. The variables are as follows:
}
\keyword{datasets}
