% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{import_countrydat}
\alias{import_countrydat}
\title{Imports a CountryDat.csv file}
\usage{
import_countrydat(
  filename = "MOZ_CountryDat.csv",
  removespace = TRUE,
  lowercase = TRUE,
  header = TRUE,
  setting = "setting"
)
}
\arguments{
\item{filename}{File name of csv.}

\item{removespace}{Removes spaces. Boolean.}

\item{lowercase}{Names in lower case. Boolean.}

\item{header}{File has column names. Boolean.}

\item{setting}{Name of 'setting' variable. String.}
}
\description{
Imports csv file in correct format (removes spaces and sets case to lower)
}
