% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{listCache}
\alias{listCache}
\alias{list_cache}
\title{List objects and values in cache.}
\usage{
listCache(simple = FALSE)

list_cache(simple = FALSE)
}
\arguments{
\item{simple}{If TRUE simply list the names of the objects in the cache,
else (FALSE) also show values.}
}
\description{
List objects and values in cache.
}
