% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{loadExperiment}
\alias{loadExperiment}
\title{Load cached data from experiment}
\usage{
loadExperiment(path)
}
\arguments{
\item{path}{Path of the experiment's folder}
}
\description{
Load cached data from experiment
}
