% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_entomology_define.R
\name{make_ento_compat}
\alias{make_ento_compat}
\title{Write entomology}
\usage{
make_ento_compat(
  baseList,
  mosqs,
  contrib = c(0.81, 0.08, 0.09, 0.02),
  propInfected = 0.078,
  propInfectious = 0.021,
  EIR = "@EIR@",
  seasonality = paste0("@m", 1:12, "@")
)
}
\arguments{
\item{baseList}{List with experiment data.}

\item{mosqs}{Mosquito vectors}

\item{contrib}{Proportion contribution}

\item{propInfected}{Proportion infected}

\item{propInfectious}{Proportion infectious}

\item{EIR}{EIR value}

\item{seasonality}{Vector of length 12}
}
\description{
Write entomology
}
