% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{make_mosquito_name}
\alias{make_mosquito_name}
\title{Generate mosquito names}
\usage{
make_mosquito_name(mosq = c("gambiaess", "funestuss"), inout = TRUE)
}
\arguments{
\item{mosq}{Name of Anopheles mosquito species (format: i.e. 'gambiaess')}

\item{inout}{If TRUE indoor and outdoor names are generated}
}
\description{
Valid names begin with 'gam' (Gambiae), 'fun' (Funestus), 'ara'
(Arabiensis) and 'alb' (Albimanus)
}
