% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_monitoring.R
\name{monitoringCohortsGen}
\alias{monitoringCohortsGen}
\title{Generate list for 'monitoring/Cohorts'}
\usage{
monitoringCohortsGen(ids)
}
\arguments{
\item{ids}{Vector containg cohort names}
}
\description{
Generates cohort sub-population entries and caches the cohort
identifiers so \code{createBaseXml()} can build the cohort-aware
\code{thirdDimension} table for postprocessing.
}
