% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_monitoring.R
\name{monitoringContinuousGen}
\alias{monitoringContinuousGen}
\title{Generate list for 'monitoring/continuous/options'}
\usage{
monitoringContinuousGen(period, options)
}
\arguments{
\item{period}{Value for period}

\item{options}{List of options}
}
\description{
Generate list for 'monitoring/continuous/options'
}
