% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_monitoring.R
\name{monitoringSurveyOptionsGen}
\alias{monitoringSurveyOptionsGen}
\title{Generate list for 'monitoring/continous/options'}
\usage{
monitoringSurveyOptionsGen(onlyNewEpisodes = NULL, options)
}
\arguments{
\item{onlyNewEpisodes}{TODO}

\item{options}{TODO}
}
\value{
List for xml contruction
}
\description{
Generate list for 'monitoring/continous/options'
}
