% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_monitoring.R
\name{monitoringSurveyTimesGen}
\alias{monitoringSurveyTimesGen}
\title{Generate list for 'monitoring/surveys/surveyTime'}
\usage{
monitoringSurveyTimesGen(
  startDate = NULL,
  endDate = NULL,
  interval,
  simStart = NULL,
  detectionLimit = NULL,
  diagnostic = NULL,
  compatSurveys = FALSE
)
}
\arguments{
\item{startDate}{Start date as character "YYYY-MM-DD".}

\item{endDate}{End date as character "YYYY-MM-DD".}

\item{interval}{Time interval. Either a string ("daily", "weekly",
"monthly", "quarterly", "yearly"), "X days" (e.g. "15 days"), or a list,
e.g. (days = c(2, 5), months = c(3:7), years = c(2005:2030)). Setting the
days in the list to 31 will always use the last day of the corresponding
month (e.g. 28 for February, 31 for July), or a vector of survey dates}

\item{simStart}{Start date of the simulation. A good idea is to put this 100
years in before your first survey date.}

\item{detectionLimit}{Deprecated in openMalaria. Double, limit above which a
human's infection is reported as patent.}

\item{diagnostic}{Name of a parameterised diagnostic to use in surveys. See
openMalaria documentation.}

\item{compatSurveys}{If survey times should be compatible to legacy
versions. By default, surveys exclude the timestep of the survey date.
Setting this to TRUE, one timestep will be added to each survey date so
they include the dates of the survey itself.}
}
\description{
Generate list for 'monitoring/surveys/surveyTime'
}
