% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MOZ}
\alias{MOZ}
\title{Demographic data from Mozambique}
\format{
An object of class \code{data.frame} with 17 rows and 2 columns.
}
\usage{
MOZ
}
\description{
A dataset with demographic data from Mozambique. The variables are as
follows:
}
\keyword{datasets}
