% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{omOutputDict}
\alias{omOutputDict}
\title{Open Malaria output dictionary}
\usage{
omOutputDict()
}
\description{
A dictionary which provides translations for the following
outputs of Open Malaria:
\itemize{
\item Survey measure numbers to names
\item If a measure's output can be separated by age group, cohort, genotype,
vector species or drug ID
\item Whether measures are aggregated between survey dates or if their values
is a snapshot
}

See
https://github.com/SwissTPH/openmalaria/wiki/MonitoringOptions#survey-measures
}
