% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{processFile}
\alias{processFile}
\title{Read file line by line}
\usage{
processFile(f, trim = TRUE, rmdups = TRUE)
}
\arguments{
\item{f}{Path to file.}

\item{trim}{If TRUE, trim whitespace.}

\item{rmdups}{If TRUE, only keep lines which are not identical to the
previous line AND which are not whitespace.}
}
\value{
Character vector with each element corresponding to a line.
}
\description{
Read file line by line and optionally remove whitespace and
duplicate lines. WARNING: Be careful with large files!
}
