% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{readOutputFile}
\alias{readOutputFile}
\title{Read Open Malaria output file}
\usage{
readOutputFile(f, translate = TRUE)
}
\arguments{
\item{f}{File name to read from.}

\item{translate}{Can be any of c("dates", "measures") or simply TRUE
(implies all) or FALSE (implies none).}
}
\description{
Read a '*_out.txt' file, apply modifications and return a data
frame which should be added to the DB.
}
