% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{readResults}
\alias{readResults}
\title{Collect Open Malaria results into a database}
\usage{
readResults(expDir, dbName, dbDir = NULL, replace = FALSE)
}
\arguments{
\item{expDir}{Database connection.}

\item{dbName}{Name of the database file without extension.}

\item{dbDir}{Directory of the database file. Defaults to the root directory.}

\item{replace}{How to handle duplicate experiments in the database. If TRUE,
any experiment with the same name will be replaced. If FALSE, a new entry
with the same name will be ignored.}
}
\description{
Collect Open Malaria results into a database
}
