% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{readScenarios}
\alias{readScenarios}
\alias{read_scenarios}
\title{Return cached scenarios}
\usage{
readScenarios(experimentDir = NULL)

read_scenarios(experimentDir = NULL)
}
\arguments{
\item{experimentDir}{Directory of the experiment}
}
\description{
Return cached scenarios
}
