% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess-new.R
\name{readSurveyOutput}
\alias{readSurveyOutput}
\title{Read survey output file}
\usage{
readSurveyOutput(outputFile, filterBySurveyMeasure = NULL)
}
\arguments{
\item{outputFile}{Survey output data file location}

\item{filterBySurveyMeasure}{Character vector of survey measures to filter
while reading from file}
}
\description{
Read survey output file
}
