% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{recXML}
\alias{recXML}
\title{TODO}
\usage{
recXML(x, data, errCol, recLevel = list())
}
\arguments{
\item{x}{TODO}

\item{data}{List with experiment information}

\item{errCol}{TODO}

\item{recLevel}{TODO}
}
\value{

}
\description{
TODO
}
