% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{run_script}
\alias{run_script}
\title{run_script is a wrapper function that runs other function with pre-defined options}
\usage{
run_script(
  scripts = c("slurm_scenarios.sh", "slurm_simulation.sh", "slurm_postprocess.sh",
    "slurm_cleanup.sh"),
  submit = TRUE,
  ...
)
}
\arguments{
\item{scripts}{specifies the process to run. Options are: "all", "scen", "malaria", "post", "jag", "sub", "clean", "fit"}

\item{submit}{if TRUE, then submits the script}

\item{...}{not used}
}
\description{
run_script is a wrapper function that runs other function with pre-defined options
}
\note{
this can run multiple scripts in a dependency
}
