% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{runScenarios}
\alias{runScenarios}
\title{Run openMalaria scenarios}
\usage{
runScenarios(scenariosDir = NULL, cmd = "openMalaria", dryRun = FALSE)
}
\arguments{
\item{scenariosDir}{Directory containing the scenario xml files. Defaults to
the cached scenario directory.}

\item{cmd}{Command to run openMalaria.}

\item{dryRun}{If TRUE, only the final command is printed but not executed.}
}
\description{
Generates a system call to launch OpenMalaria with the correct
parameters for each scenario found in the directory. It is assumed that
OpenMalaria is available in the PATH.
}
