% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{runSimulations}
\alias{runSimulations}
\alias{run_simulations}
\title{Run openMalaria scenarios}
\usage{
runSimulations(
  scenarios = NULL,
  cmd = "openMalaria",
  dryRun = FALSE,
  verbose = FALSE,
  ncores = 1,
  rowStart = NULL,
  rowEnd = NULL
)

run_simulations(
  scenarios = NULL,
  cmd = "openMalaria",
  dryRun = FALSE,
  verbose = FALSE,
  ncores = 1,
  rowStart = NULL,
  rowEnd = NULL
)
}
\arguments{
\item{scenarios}{Scenario data frame}

\item{cmd}{Command to run openMalaria.}

\item{dryRun}{If TRUE, only the final command is printed but not executed.}

\item{verbose}{If TRUE, use OpenMalaria's verbose output.}

\item{ncores}{Number of parallel processes to use.}

\item{rowStart}{Starting row. Optional.}

\item{rowEnd}{End row. Optional.}
}
\description{
Generates a system call to launch OpenMalaria with the correct
parameters for each scenario found in the directory. It is assumed that
OpenMalaria is available in the PATH.
}
