% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directories.R
\name{setupDirs}
\alias{setupDirs}
\alias{setup_dirs}
\title{Create folder structure}
\usage{
setupDirs(
  experimentName,
  rootDir = NULL,
  scenariosDir = NULL,
  logsDir = NULL,
  replace = "ask"
)

setup_dirs(
  experimentName,
  rootDir = NULL,
  scenariosDir = NULL,
  logsDir = NULL,
  replace = "ask"
)
}
\arguments{
\item{experimentName}{Name of the experiment. Spaces will be replaced by
underscores.}

\item{rootDir}{Root directory as a string. Defaults to the current working
directory.}

\item{scenariosDir}{Scenarios directory name as a string. Will be a
subfolder of the \code{rootDir}. Defaults to 'scenarios'.}

\item{logsDir}{Logs directory name as a string. Will be a subfolder of the
\code{rootDir}. Defaults to 'logs'.}

\item{replace}{If the present directory structure should be replaced. Can be
TRUE, FALSE or "ask". This will remove any existing content!}
}
\description{
Generate the folder structure for an experiment.
}
