% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{setupOM}
\alias{setupOM}
\alias{setup_om}
\title{Download required Open Malaria files}
\usage{
setupOM(version = 43, dir = NULL)

setup_om(version = 43, dir = NULL)
}
\arguments{
\item{version}{Major schema version. Supported: 43}

\item{dir}{Target directory. Defaults to experiment directory.}
}
\description{
Download required Open Malaria files
}
