% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm_postprocessing.R
\name{slurmPrepareCleanup}
\alias{slurmPrepareCleanup}
\title{Run preparations for SLURM submission of postprocessing cleanup}
\usage{
slurmPrepareCleanup(
  expName,
  scenarios,
  basename,
  files_per_job = 200,
  seed_as_hist_param = TRUE,
  create_monthly_dataset = FALSE,
  mem_clean = "2GB",
  time_clean = "00:10:00",
  qos_clean = "30min",
  setting = "setting"
)
}
\arguments{
\item{expName}{Name of experiment}

\item{scenarios}{Scenario data frame}

\item{basename}{Name of base xml file}

\item{files_per_job}{Number of scenario files to have in each dataset (too
many, and it will be slow)}

\item{seed_as_hist_param}{If true, then each seed is a different historical
parameter}

\item{create_monthly_dataset}{If true, creates a CombinedDat_month, where
each month and outcome is a column}

\item{mem_clean}{Memory for process script}

\item{time_clean}{Time for process script}

\item{qos_clean}{Queue for process script}

\item{setting}{"setting"}
}
\description{
Run preparations for SLURM submission of postprocessing cleanup
}
