% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm_postprocessing.R
\name{slurmPreparePostprocess}
\alias{slurmPreparePostprocess}
\title{Run preparations for SLURM submission of postprocess}
\usage{
slurmPreparePostprocess(
  expName,
  scenarios,
  basename,
  variables_of_interest = c("PR", "nUncomp", "nHost", "nSevere", "incidence", "ddeath",
    "edeath", "edirdeath", "expectedDirectDeaths", "nTreatments1", "nTreatments2",
    "nTreatments3", "nHospitalSeqs", "nHospitalRecovs", "expectedHospitalDeaths",
    "tUncomp", "tSevere"),
  age_groups_of_interest = c("2to10", "0to5", "All"),
  files_per_job = 200,
  seed_as_hist_param = TRUE,
  create_monthly_dataset = FALSE,
  variable_in_monthly_dataset = "PR",
  years_in_monthly_dataset = 2015:2020,
  mem_process = "2GB",
  time_process = "00:06:00",
  qos_process = "6hours",
  ignore_these_variables = c("propOut", "jan", "feb", "mar", "apr", "may", "jun",
    "jul", "aug", "sep", "oct", "nov", "dec", "scaled_down_flag"),
  setting = "setting"
)
}
\arguments{
\item{expName}{Name of experiment}

\item{scenarios}{Scenario data frame}

\item{basename}{Name of base xml file}

\item{variables_of_interest}{Variables to store in the dataset}

\item{age_groups_of_interest}{Age groups to aggregate to}

\item{files_per_job}{Number of scenario files to have in each dataset (too
many, and it will be slow)}

\item{seed_as_hist_param}{If true, then each seed is a different historical
parameter}

\item{create_monthly_dataset}{If true, creates a CombinedDat_month, where
each month and outcome is a column}

\item{variable_in_monthly_dataset}{(e.g. "PR" or "incidence")}

\item{years_in_monthly_dataset}{Which years to store data for}

\item{mem_process}{Memory for process script}

\item{time_process}{Time for process script}

\item{qos_process}{Queue for process script}

\item{ignore_these_variables}{Variables to ignore in post-processing}

\item{setting}{"setting"}
}
\description{
Run preparations for SLURM submission of postprocess
}
