% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm.R
\name{splitSeq}
\alias{splitSeq}
\title{Divide a sequence x into chunks of size n. Any rest is appended.}
\usage{
splitSeq(x, n)
}
\arguments{
\item{x}{Sequence like 1:10}

\item{n}{Chunk size}
}
\description{
Divide a sequence x into chunks of size n. Any rest is appended.
}
