% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{storeScenarios}
\alias{storeScenarios}
\alias{store_scenarios}
\title{Store scenarios in cache folder}
\usage{
storeScenarios(scenarios, csv = TRUE)

store_scenarios(scenarios, csv = TRUE)
}
\arguments{
\item{scenarios}{Data frame containing the values for the placeholders. One
row per scenario, placeholders in columns. Column names correspond to the
placeholder names.}

\item{csv}{Additionally save scenarios as .csv file.}
}
\description{
Store scenarios in cache folder
}
