% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_monitoring.R
\name{surveyAgeGroupsGen}
\alias{surveyAgeGroupsGen}
\title{Generate list for 'monitoring/ageGroup'}
\usage{
surveyAgeGroupsGen(lowerbound, upperbounds)
}
\arguments{
\item{lowerbound}{Double, lower bound of age group}

\item{upperbounds}{Vector with upperbounds}
}
\value{
List for xml contruction
}
\description{
Generate list for 'monitoring/ageGroup'
}
