% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{syncCache}
\alias{syncCache}
\alias{sync_cache}
\title{Synchronized cached data from experiment}
\usage{
syncCache(path)

sync_cache(path)
}
\arguments{
\item{path}{Path of the experiment's folder}
}
\description{
Synchronized cached data from experiment
}
