% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing.R
\name{toggleDebug}
\alias{toggleDebug}
\title{Toggle debug messages}
\usage{
toggleDebug()
}
\description{
Enables the output of debug messages in functions which have
that implemented. Debug output is off by default.
}
