% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing.R
\name{toggleVerbose}
\alias{toggleVerbose}
\title{Toggle verbose messages}
\usage{
toggleVerbose()
}
\description{
Enables the output of verbose messages in functions which have
that implemented. Verbose output is off by default.
}
