% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UGA}
\alias{UGA}
\title{Demographic data from Uganda}
\format{
An object of class \code{data.frame} with 20 rows and 2 columns.
}
\usage{
UGA
}
\description{
A dataset with demographic data from Uganda. The variables are as follows:
}
\keyword{datasets}
