% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_validation.R
\name{validateXML}
\alias{validateXML}
\alias{validate_xml}
\title{Validate the XML file}
\usage{
validateXML(xmlfile = NULL, schema = NULL, scenarios = NULL)

validate_xml(xmlfile = NULL, schema = NULL, scenarios = NULL)
}
\arguments{
\item{xmlfile}{Path to XML file.}

\item{schema}{Path to XSD schema file.}

\item{scenarios}{Scenario data frame.}
}
\description{
Validate the XML file
}
