% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{convert_access}
\alias{convert_access}
\title{Function to scale from care seeking to effective treatment coverage per 14
days to 5 days}
\usage{
convert_access(
  dat,
  pattern = "Access",
  katya = T,
  country = NULL,
  scale = NULL
)
}
\arguments{
\item{dat}{Country data}

\item{pattern}{How to identify 'Access' variables}

\item{katya}{If true, then scales from care seeking to effective treatment
coverage}

\item{country}{Abbreviation (alpha-3 code ISO 3166) if using a country
pre-specified scaling factor}

\item{scale}{Scaling factor, if NULL, a default will be used based on
Galactionova et al 2012}
}
\description{
Function to scale from care seeking to effective treatment coverage per 14
days to 5 days
}
\examples{
# Input dataset
dat <- data.frame(setting = "alpha", access2005 = .5, access2006 = .1, access2010 = .9)
# Converting access values
dat <- convert_access(dat = dat, pattern = "access", country = "MOZ")
# Visualizing the dataset
# view_past(dat = dat, pattern = "access")
}
