% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_demography.R
\name{write_demography_compat}
\alias{write_demography_compat}
\title{Write xml chunk for demography}
\usage{
write_demography_compat(
  baseList,
  maxage = 90,
  country = "BEN",
  percents = NULL,
  uppers = NULL,
  pop = "@pop@"
)
}
\arguments{
\item{baseList}{List with experiment data.}

\item{maxage}{Maximum age of the human population}

\item{country}{Country (abbreviation, format "BEN")}

\item{percents}{Percents}

\item{uppers}{Upper age limit}

\item{pop}{Population size in simulations}
}
\description{
Write xml chunk for demography
}
