% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_model.R
\name{write_end_compat}
\alias{write_end_compat}
\title{Write the underlying transmission dynamics}
\usage{
write_end_compat(
  baseList,
  seed = 4,
  modelname = "base",
  selectedOptions = NULL,
  comorb = 0.0968,
  critage = 0.117383,
  logodds = 0.736202,
  healthSystemMemory = "6t",
  ...
)
}
\arguments{
\item{baseList}{List with experiment data.}

\item{seed}{Number of replicates per scenario.}

\item{modelname}{Select model (default= base 'R0000'). Other options
currently not available.}

\item{selectedOptions}{Model options to activate. Can be any combination of
c("COMORB_HET", "VECTOR_SIMPLE_MPD_MODEL", "LOGNORMAL_MASS_ACTION",
"MAX_DENS_CORRECTION", "INNATE_MAX_DENS", "INDIRECT_MORTALITY_FIX").}

\item{comorb}{Comorbidity.}

\item{critage}{Critical age.}

\item{logodds}{Log odds terms.}

\item{healthSystemMemory}{healthSystemMemory (see OpenMalaria).}

\item{...}{deprecated}
}
\description{
Write the underlying transmission dynamics
}
