% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_healthSystem.R
\name{write_healthsys_compat}
\alias{write_healthsys_compat}
\title{Writes the health system part of the xml file}
\usage{
write_healthsys_compat(
  baseList,
  access = "@Access2000@",
  init.a = 1,
  init.q = 1,
  init.s = 1,
  comp.a = 1,
  comp.q = 1,
  comp.s = 1,
  pSeekOfficialCareSevere = 0.48,
  pSelfTreatUncomplicated = 0.01821375
)
}
\arguments{
\item{baseList}{List with experiment data.}

\item{access}{Defines the baseline effective treatment coverage}

\item{init.a}{ACT initial value}

\item{init.q}{QN initial value}

\item{init.s}{self-treat initial value}

\item{comp.a}{Compliance value ACT}

\item{comp.q}{Compliance value QN}

\item{comp.s}{Compliance value self-treat}

\item{pSeekOfficialCareSevere}{Default value is 0.48}

\item{pSelfTreatUncomplicated}{Default value is 0.0182}
}
\description{
Writes the health system part of the xml file
}
