% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_monitoring.R
\name{write_monitoring_compat}
\alias{write_monitoring_compat}
\title{Writes the monitoring measures xml chunk}
\usage{
write_monitoring_compat(
  baseList,
  name = "Annual Surveys",
  continuous = TRUE,
  SurveyOptions = TRUE,
  surveyMeasures = NULL,
  continuousMeasures = NULL,
  y1 = 2000,
  y2 = 2036,
  m1 = 1,
  m2 = 1,
  d1 = 1,
  d2 = 1,
  SIMSTART = "1918-01-01",
  interval = "month",
  detect = 200,
  upperbounds = c(1, 2, 5, 6, 10, 11, 100)
)
}
\arguments{
\item{baseList}{List with experiment data.}

\item{name}{Name of monitoring settings.}

\item{continuous}{Write measures for the continuous surveys}

\item{SurveyOptions}{Write measures for the SurveyOptions}

\item{surveyMeasures}{If NULL, default will be used}

\item{continuousMeasures}{If NULL, default will be used}

\item{y1}{Year of the first date (surveys starting from year y1)}

\item{y2}{Year of the end date (surveys continuing until year y2)}

\item{m1}{Month of the first date}

\item{m2}{Month of the end date}

\item{d1}{Day of the first date}

\item{d2}{Day of the end date}

\item{SIMSTART}{Starting date of the simulations in the format "yyyy-mm-dd",
default e.g. "1918-01-01"}

\item{interval}{Interval "month" "year" "quarter"}

\item{detect}{Detection limit for parasites in the blood}

\item{upperbounds}{Defines the upper age bound per age group}
}
\description{
Wrapper function for write_continuous and write_SurveyOptions
}
