% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{write_scen_data}
\alias{write_scen_data}
\title{Description: Function to write scenario data}
\usage{
write_scen_data(scens, full, nameExperiment, startnum = 1, saveit = TRUE, ...)
}
\arguments{
\item{scens}{Dataset of scenarios}

\item{full}{List of experiment variables and values}

\item{nameExperiment}{Name of the experiment (string)}

\item{startnum}{Start number for numbering scenarios, default = 1}

\item{saveit}{If TRUE, then saves the scens.RData object}

\item{...}{Deprecated options}
}
\description{
Description: Function to write scenario data
}
