% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{writeCache}
\alias{writeCache}
\alias{write_cache}
\title{Write cached data from experiment}
\usage{
writeCache(path)

write_cache(path)
}
\arguments{
\item{path}{Path of the experiment's folder}
}
\description{
Write cached data from experiment
}
