% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{xmlTimeGen}
\alias{xmlTimeGen}
\title{Generate date sequence based on start and end date}
\usage{
xmlTimeGen(startDate, endDate, interval)
}
\arguments{
\item{startDate}{TODO}

\item{endDate}{TODO}

\item{interval}{TODO}
}
\description{
Generate date sequence based on start and end date
}
