% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_interventions_changeHS.R
\name{changeHSpSeqInGen}
\alias{changeHSpSeqInGen}
\title{Generate list for 'changeHS/SpSeq'}
\usage{
changeHSpSeqInGen(interpolation = NULL, ageGroups)
}
\arguments{
\item{interpolation}{Value for interpolation}

\item{ageGroups}{Data frame}
}
\description{
Generate list for 'changeHS/SpSeq'
}
