% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{cluster_status}
\alias{cluster_status}
\title{writes a file of all job status and writes a table showing categories (failed, completed)}
\usage{
cluster_status(
  AnalysisDir = getwd(),
  username = as.character((Sys.info()["user"])),
  stime = "00:00:00",
  sdate = Sys.Date()
)
}
\arguments{
\item{AnalysisDir}{analysis directory folder}

\item{username}{slurm username}

\item{stime}{start time 00:00:00 for checking jobs ("hh:mm:ss")}

\item{sdate}{start date for checking jobs "yyyy-mm-dd"}
}
\value{
table of status of jobs on cluster
}
\description{
writes a file of all job status and writes a table showing categories (failed, completed)
}
