% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{createBaseXml}
\alias{createBaseXml}
\title{Create an openMalaria experiment}
\usage{
createBaseXml(data = NULL, env = parent.frame())
}
\arguments{
\item{data}{List containing all information}
}
\value{

}
\description{
Processes a list as containing all information to generate all required
folders and files for openMalaria. Spaces in the name of the experiment are
automatically replaced by underscores.
}
