% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_demography.R
\name{defineDemography}
\alias{defineDemography}
\title{Define and write demography input to baseList}
\usage{
defineDemography(
  baseList,
  demographyData = NULL,
  popSize = 3000,
  upperbound = NULL,
  poppercent = NULL,
  lowerbound = NULL,
  name = NULL,
  maximumAgeYrs = NULL
)
}
\arguments{
\item{baseList}{List with experiment data.}

\item{demographyData}{List with demography data in specific format}

\item{popSize}{Number of agents in openMalaria}

\item{upperbound}{Upper bound of age group}

\item{poppercent}{Percentage of human population in age group}

\item{lowerbound}{Lower bound of age group}

\item{name}{Name of demography data}

\item{maximumAgeYrs}{Maximum age of simulated humans in years}
}
\description{
Define and write demography input to baseList
}
