% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_entomology_define.R
\name{defineEntomology}
\alias{defineEntomology}
\alias{define_entomology}
\title{Defines entomology setting}
\usage{
defineEntomology(baseList, mosquito_parameters, append = TRUE)

define_entomology(baseList, mosquito_parameters, append = TRUE)
}
\arguments{
\item{baseList}{List with experiment data.}

\item{mosquito_parameters}{Mosquito bionomics parameterization list of
different mosquito species as obtained from AnophelesModel package
function get_OM_ento_snippet}

\item{append}{If TRUE, then append to existing baseList, otherwise overwrite}
}
\description{
Defines entomology setting
}
