% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_interventions_define.R
\name{defineVectorControl}
\alias{defineVectorControl}
\alias{define_vector_control}
\title{Adds vector control intervention parameterisation to baseList}
\usage{
defineVectorControl(
  baseList,
  intervention_parameterization,
  append = TRUE,
  name = NULL,
  hist = FALSE,
  resistance = 0.1
)

define_vector_control(
  baseList,
  intervention_parameterization,
  append = TRUE,
  name = NULL,
  hist = FALSE,
  resistance = 0.1
)
}
\arguments{
\item{baseList}{List with experiment data.}

\item{intervention_parameterization}{Vector control intervention
parameterization list depending on three parameters (deterrency,
preprandrial, postprandial) and decay functions:}

\item{append}{If TRUE, then append to existing baseList, otherwise
overwrites}

\item{name}{Name tag list}

\item{hist}{If TRUE, then decay is assumed to be step function set to 1 for
a year and then to zero for the remainder}

\item{resistance}{Scaling function of insecticide resistance}
}
\description{
Adds vector control intervention parameterisation to baseList
}
