% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_interventions_deploy.R
\name{deployIT}
\alias{deployIT}
\alias{deploy_IT}
\title{Writes the deployment of an intervention}
\usage{
deployIT(
  baseList,
  component = "ITN",
  cumulative = FALSE,
  effects = NULL,
  startDate = NULL,
  endDate = NULL,
  interval,
  minAge = NULL,
  maxAge = NULL,
  coverage = NULL,
  subpop = FALSE
)

deploy_IT(
  baseList,
  component = "ITN",
  cumulative = FALSE,
  effects = NULL,
  startDate = NULL,
  endDate = NULL,
  interval,
  minAge = NULL,
  maxAge = NULL,
  coverage = NULL,
  subpop = FALSE
)
}
\arguments{
\item{baseList}{List with experiment data.}

\item{component}{Name of intervention.}

\item{cumulative}{default is FALSE. Do not set to TRUE.}

\item{effects}{Either NULL or c("det","pre","post")}

\item{startDate}{Date in YYYY-MM-DD format.}

\item{endDate}{Date in YYYY-MM-DD format.}

\item{interval}{A string like '1 weeks'. Same as in \code{\link[=seq.Date]{seq.Date()}}. Or a list
composed of the entries 'days' (optional), 'months' (optional) and
'years'. If a list is used, startDate and endDate are not used and can be
NULL.}

\item{minAge}{Minimum age for deployment (used in SMC)}

\item{maxAge}{Maximum age for deployment (used in SMC)}

\item{coverage}{Value or variable of coverage}

\item{subpop}{If TRUE, then restricts to a subpopulation (see
restrictToSubPop in OpenMalaria)}
}
\description{
Writes the deployment of an intervention
}
