% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{.createTables}
\alias{.createTables}
\title{Create the DB table layout}
\usage{
.createTables(connection)
}
\arguments{
\item{connection}{Database connection.}
}
\description{
Creates the database schema: experiments, scenarios,
scenarios_metadata, placeholders, results. Changes are cascaded from down
from experiments table.
}
\keyword{internal}
