% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess-new.R
\name{.numberToSurveyMeasure}
\alias{.numberToSurveyMeasure}
\title{@title Dictionary of survey measure numbers to names}
\usage{
.numberToSurveyMeasure()
}
\description{
@title Dictionary of survey measure numbers to names
}
\keyword{internal}
