% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{.placeholderCollect}
\alias{.placeholderCollect}
\title{Collect if current value is a placeholder}
\usage{
.placeholderCollect(x)
}
\arguments{
\item{x}{A R object.}
}
\description{
Collect if current value is a placeholder
}
