% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_interventions_deploy.R
\name{.placeholderseqGen}
\alias{.placeholderseqGen}
\title{Generate placeholder sequence}
\usage{
.placeholderseqGen(x, placeholders)
}
\arguments{
\item{x}{Input list. First element has to a name, second a sequence, e.g.
list("name", c(seq1, seq2, ...))}

\item{placeholders}{Character vector of names which should be processed}
}
\description{
Generate placeholder sequence
}
\keyword{internal}
