% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_interventions_deploy.R
\name{.placeholderseqGen}
\alias{.placeholderseqGen}
\title{Generate placeholder sequence}
\usage{
.placeholderseqGen(x, placeholders)
}
\arguments{
\item{x}{Input list. It is assumed that this list is generated from a
'as.list(match.call())' call and thus contains unevaluated expressions. A
similar input can be generated by something like 'list(foo = call("list",
"foo", c(1, 2, 3, ...)))'.}

\item{placeholders}{Character vector of names which should be processed}
}
\description{
Generate placeholder sequence
}
\keyword{internal}
