% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{.surveyMeasuresDict}
\alias{.surveyMeasuresDict}
\title{Returns a list, mapping measure numbers and names}
\usage{
.surveyMeasuresDict()
}
\description{
Returns a list, mapping measure numbers and names
}
