% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{.xmlAddChunks}
\alias{.xmlAddChunks}
\title{Generate XML chunks}
\usage{
.xmlAddChunks(outlist, element, attributeList)
}
\arguments{
\item{outlist}{List which will get modified and returned.}

\item{element}{Element to add}

\item{attributeList}{Attributes to add}
}
\value{
XML entries
}
\description{
Generate XML chunks
}
\keyword{internal}
