% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{.xmlMakeDocRec}
\alias{.xmlMakeDocRec}
\title{Create xml entries recursively}
\usage{
.xmlMakeDocRec(baseXML, data)
}
\arguments{
\item{baseXML}{The root xml object}

\item{data}{List with experiment information}
}
\description{
Create xml entries recursively
}
\keyword{internal}
