% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{loadExperiment}
\alias{loadExperiment}
\alias{load_experiment}
\alias{loadCache}
\alias{load_cache}
\title{Load cached data from experiment}
\usage{
loadExperiment(path)

load_experiment(path)

loadCache(path)

load_cache(path)
}
\arguments{
\item{path}{Path of the experiment's folder}
}
\description{
Load cached data from experiment
}
