% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_monitoring.R
\name{monitoringCohortsGen}
\alias{monitoringCohortsGen}
\title{Generate list for 'monitoring/Cohorts'}
\usage{
monitoringCohortsGen(ids)
}
\arguments{
\item{ids}{Vector containg cohort names}
}
\description{
Generate list for 'monitoring/Cohorts'
}
