% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_monitoring.R
\name{monitoringContinousGen}
\alias{monitoringContinousGen}
\title{Generate list for 'monitoring/continous/options'}
\usage{
monitoringContinousGen(period, options)
}
\arguments{
\item{period}{TODO}

\item{options}{TODO}
}
\value{
List for xml contruction
}
\description{
Generate list for 'monitoring/continous/options'
}
