% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_monitoring.R
\name{monitoringSurveyOptionsGen}
\alias{monitoringSurveyOptionsGen}
\title{Generate list for 'monitoring/continuous/options'}
\usage{
monitoringSurveyOptionsGen(onlyNewEpisodes = NULL, options)
}
\arguments{
\item{onlyNewEpisodes}{Value for onlyNewEpisodes}

\item{options}{List of options}
}
\description{
Generate list for 'monitoring/continuous/options'
}
