% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_monitoring.R
\name{monitoringSurveyTimesGen}
\alias{monitoringSurveyTimesGen}
\title{Generate list for 'monitoring/surveys/surveyTime'}
\usage{
monitoringSurveyTimesGen(
  startDate,
  endDate,
  interval,
  detectionLimit = NULL,
  diagnostic = NULL
)
}
\arguments{
\item{startDate}{Start date as character "YYYY-MM-DD".}

\item{endDate}{End date as character "YYYY-MM-DD".}

\item{interval}{Time interval. Either a string, e.g. "1 week", or a list,
e.g. (days = c(2, 5), months = c(3:7), years = c(2005:2030)). Accepted
string are day(s), week(s), month(s) and quarter(s). Setting the days in
the list to 31 will always use the last day of the corresponding month
(e.g. 28 for February, 31 for July).}

\item{detectionLimit}{Deprecated in openMalaria. Double, limit above which a
human's infection is reported as patent.}

\item{diagnostic}{Name of a parameterised diagnostic to use in surveys. See
openMalaria documentation.}
}
\description{
Generate list for 'monitoring/surveys/surveyTime'
}
