% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{omOutputDict}
\alias{omOutputDict}
\title{Open Malaria output dictionary}
\usage{
omOutputDict()
}
\description{
A dictionary which provides tranlations for the following
outputs of Open Malaria:
\itemize{
\item Survey measure numbers to names
\item Whether measures are summed up between survey dates (incident = TRUE) or
represent prevalent characteristics (incident = FALSE)
\item An identifier for the 'third dimension' column. This can be 'age_group',
'vector_species', 'drug_id' or NA
}

See: https://github.com/SwissTPH/openmalaria/wiki/MonitoringOptions
}
