% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{putCache}
\alias{putCache}
\title{Store object in cache.}
\usage{
putCache(x, value)
}
\arguments{
\item{x}{Object name to store in cache.}

\item{value}{Object value to store in cache.}
}
\description{
Store object in cache.
}
