% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{readOutputFile}
\alias{readOutputFile}
\title{Read Open Malaria output file}
\usage{
readOutputFile(f, filter = NULL, translate = TRUE, scenID = NULL)
}
\arguments{
\item{f}{File name to read from.}

\item{filter}{A data.table expression to filter rows.}

\item{translate}{Can be any of c("dates", "measures", "third_dimension") or
simply TRUE (implies all) or FALSE (implies none).}

\item{scenID}{Scenario ID to be added.}
}
\description{
Read a '*_out.txt' file, apply modifications and return a data
frame which should be added to the DB.
}
