% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{recXML}
\alias{recXML}
\title{Recursively create XML}
\usage{
recXML(x, data, errCol, recLevel = list())
}
\arguments{
\item{x}{XML node}

\item{data}{List with experiment information}

\item{errCol}{Checkmate error collection}

\item{recLevel}{List containing current recursion level}
}
\description{
Recursively create XML
}
\keyword{internal}
