% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{runScenarios}
\alias{runScenarios}
\title{Run openMalaria scenarios}
\usage{
runScenarios(
  scenarios,
  cmd = "openMalaria",
  dryRun = FALSE,
  verbose = FALSE,
  ncores = 1,
  rowStart = NULL,
  rowEnd = NULL
)
}
\arguments{
\item{cmd}{Command to run openMalaria.}

\item{dryRun}{If TRUE, only the final command is printed but not executed.}

\item{verbose}{If TRUE, use OpenMalaria's verbose output. This option}

\item{scenariosDir}{Directory containing the scenario xml files. Defaults to
the cached scenario directory.}
}
\description{
Generates a system call to launch OpenMalaria with the correct
parameters for each scenario found in the directory. It is assumed that
OpenMalaria is available in the PATH.
}
