% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{runScenarios}
\alias{runScenarios}
\title{Run openMalaria scenarios}
\usage{
runScenarios(scenariosDir = NULL, cmd = "openMalaria")
}
\arguments{
\item{scenariosDir}{Directory containing the scenario xml files. Defaults to
the cached scenario directory.}

\item{cmd}{Command to run openMalaria.}
}
\description{
Run openMalaria scenarios
}
