% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{runScenarios}
\alias{runScenarios}
\title{Run openMalaria scenarios}
\usage{
runScenarios(scenariosDir = omuCache$scenariosDir)
}
\arguments{
\item{scenariosDir}{Directory containing the scenario xml files. Defaults to
the cached scenario directory.}
}
\description{
Run openMalaria scenarios
}
